### 2023-03-02 #84 @Midnighter
taxpasta: TAXonomic Profile Aggregation and STAndardisation
Submitting Author: Moritz E. Beber (@Midnighter)
All current maintainers: (@Midnighter, @sofstam, @jfy133)
Package Name: taxpasta
One-Line Description of Package: TAXonomic Profile Aggregation and STAndardisation
Repository Link:  https://github.com/taxprofiler/taxpasta
Version submitted:  0.2.1
Archive: [TBD](https://github.com/taxprofiler/taxpasta/releases/tag/0.4.0)
Version accepted: 0.4.0

---

The main purpose of taxpasta is to _standardise_ taxonomic profiles created by a
range of bioinformatics tools. We call those tools taxonomic profilers. They
each come with their own particular, tabular output format. Across the profilers,
relative abundances can be reported in read counts, fractions, or percentages,
as well as any number of additional columns with extra information. We therefore
decided to take [the lessons learnt](https://xkcd.com/927/) to heart and provide
our own solution to deal with this pasticcio. With taxpasta you can ingest all
of those formats and, at a minimum, output taxonomy identifiers and their
integer counts.

Taxpasta can not only standardise profiles but also _merge_ them across samples
for the same profiler into a single table. In future, we also intend to offer
methods for forming a _consensus_ for the same sample analyzed by different
profilers.

	- [ ] Data visualization **

> ** Please fill out a pre-submission inquiry before submitting a data visualization package.*

    Taxpasta is a tool for anyone working with taxonomic profiles from metagenomic sequencing experiments. Mostly that means ecologists, bioinformaticians, statisticians. Taxpasta's main application is to standardise profiles from a range of different tools. Having a singular format facilitates downstream analyses. Taxpasta is used, for example, in the upcoming [taxprofiler pipeline](https://nf-co.re/taxprofiler) implemented in nextflow. There, it also serves to combine the profiles of many samples into a single file.

    The [BIOM format](https://biom-format.org/) was created with the intention of standardizing a storage format for microbiome analyses. However, creating this format was entirely left to the user. Taxpasta conveniently knows how to read profiles from a range of tools and can also produce BIOM output.

    Some of the taxonomic profilers also come with scripts to convert their output into another format but none of them support such a wide range of tools as taxpasta does.

- [x] contains a README with instructions for installing the development version (development version is described in `CONTRIBUTING.rst`).

[Templates]: https://www.pyopensci.org/software-peer-review/appendices/templates.html
