### 2023-03-02 bff9d3b
docs: expand about section and add to main readme
