### 2023-03-03 afceb9f
Apply suggestions from code review
Co-authored-by: James A. Fellows Yates <jfy133@gmail.com>
