+++ README.md
@@ -12,16 +12,17 @@ _TAXonomic Profile Aggregation and STAndardisation_
 
 The main purpose of taxpasta is to _standardise_ taxonomic profiles created by a
 range of bioinformatics tools. We call those tools taxonomic profilers. They
-each come with their own particular tabular output format. Relative abundances
-in read counts, fractions, or percentages. Any number of columns with additional
+each come with their own particular tabular output format. Across the profilers, relative abundances can be reported
+in read counts, fractions, or percentages, as well as any number of additional columns with extra
 information. We therefore decided to take [the lessons
 learnt](https://xkcd.com/927/) to heart and provide our own solution to deal
 with this pasticcio. With taxpasta you can ingest all of those formats and, at
 a minimum, output taxonomy identifiers and their integer counts.
 
 Taxpasta can not only standardise profiles but also _merge_ them
-across samples for the same profiler. In future, we also intend to offer methods
+across samples for the same profiler into a single table. In future, we also intend to offer methods
 for forming a _consensus_ for the same sample analyzed by different profilers.
+
 ## Install
 
 It's as simple as:
