### 2023-03-06 #84 @lwasser
hey there 👋  @Midnighter welcome to pyopensci and thank you for this submission! i just wanted to say hello so you know that we see this submission and will be getting back to you shortly! i also am looking into forms for the submission template as well - many thanks for that suggestion! More soon!
