### 2023-03-08 #84 @NickleDave
Welcome @Midnighter @sofstam @jfy133!
I'm adding initial checks for this package.

Glad to see this as someone who develops a similar tool (for a different domain!).

Looks like most everything is there.

I do want to request one thing before we proceed with a review though:
can you please add at least one full tutorial?

Something like the "Examples" section in the Pynteny docs here (which also has a CLI interface): https://robaina.github.io/Pynteny/examples/example_cli/

(These docs were in place before we began the Pynteny review, which is why I'm providing them as an example of what we need to see before we start.)

I would suggest a sort of walkthrough of the main use case for taxpasta. This might be a simplified version of something you're using it for in research already.

I have provided some other feedback below but adding an initial tutorial is the only thing that's necessary at this time.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

  - @Midnighter given the domain you're targeting, maybe it be good to publish a distribution package on bioconda and/or conda forge?
  - [x] The package imports properly into a standard Python environment `import package-name`.
  - yes this is clearly a data munging package
    - I would strongly suggest adding a more general description of the package right at the top of the index page, that even a non-generalist (like me) can understand. What do people use taxonomic profiles for in your field? Why would it help me to be able to convert between formats? Links to Wikipedia pages, review articles, etc., would be helpful here. Ideally with a brief example of a table, like those shown in the command docs ("taxpasta is an interoperability tool for working with taxonomic profiles, that look something like this:"
    - There are install instructions but no real examples on the landing/index page. At a minimum the output of `taxpasta -h` could be shown and additionally a brief snippet. If the main usage is the CLI, then perhaps demos of usage could be shown with a tool like [asciicinema](https://asciinema.org/)
      - Also the output of `taxpasta -h` was truncated for me. See below. Not sure it would be clear to a new user what `consensus` and `merge` commands do because of the ellipsis. From the docs I guess this is because your cli package imports text from somewhere else? Is there a way to not truncate? Maybe just give a briefer summary?
```
$ taxpasta -h
Usage: taxpasta [OPTIONS] COMMAND [ARGS]...

  TAXonomic Profile Aggregation and STAndardisation

Options:
  --version                       Print only the current tool version and
                                  exit.
  -l, --log-level [DEBUG|INFO|WARNING|ERROR|CRITICAL]
                                  Set the desired log level.  [default: INFO]
  --install-completion [bash|zsh|fish|powershell|pwsh]
                                  Install completion for the specified shell.
  --show-completion [bash|zsh|fish|powershell|pwsh]
                                  Show completion for the specified shell, to
                                  copy it or customize the installation.
  -h, --help                      Show this message and exit.

Commands:
  consensus    Form a consensus for the same sample but from different...
  merge        Standardise and merge two or more taxonomic profiles into...
  standardise  Standardise a taxonomic profile (alias: 'standardize').
```
    - There are some snippet-like examples in documentation for commands of the command-line interface but I do not find more extensive tutorials
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
    - @Midnighter is there a link to the CONTRIBUTING.rst in the README? I'm not finding it. If not, I'd suggest adding it.
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
    - inside the .github directory
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---

This is not required in checks but I would strongly suggest adding example data to taxpasta.
Since you are working with tabular data that I'd guess can be relatively lightweight, it should be relatively painless to add and it will really help people understand package usage.
Again see the Pynteny example where they work with example data (not suggesting you need to add a download command but just showing you why built-in data is useful: https://robaina.github.io/Pynteny/examples/example_cli/#download-pgap-profile-hmm-database)

For my own library I just add the files directly to the package and then access with `importlib.resources`, see this issue: https://github.com/vocalpy/crowsetta/issues/90
Some libraries have adopted the [`pooch` library](https://github.com/fatiando/pooch) for data, e.g. scikit-image:  https://scikit-image.org/docs/stable/api/skimage.data.html. Might be worth snooping their issues to help you figure out how to implement if you decide to adopt it.
