+++ docs/supported_profilers/krakenuniq.md
@@ -9,7 +9,7 @@ Taxpasta expects a 9 column table. This file is generated with the KrakenUniq pa
 | Column Header | Description |
 | ------------- | ----------- |
 | %             |             |
-| reads *       |             |
+| reads \*      |             |
 | taxReads      |             |
 | kmers         |             |
 | dup           |             |
@@ -18,7 +18,7 @@ Taxpasta expects a 9 column table. This file is generated with the KrakenUniq pa
 | rank          |             |
 | taxName       |             |
 
-> * Value used in standardised profile output 
+> \* Value used in standardised profile output
 
 ## Example
 
