+++ docs/supported_profilers/metaphlan.md
@@ -6,14 +6,14 @@
 
 Taxpasta expects a four-column, tab-separated file. This is generated by redirecting the MetaPhlAn3 `stdout`. It interprets the columns as:
 
-| Column Header        | Description |
-|----------------------|-------------|
-| clade_name           |             |
-| taxonomy_id          |             |
-| relative_abundance * |             |
-| additional_species   |             |
+| Column Header         | Description |
+| --------------------- | ----------- |
+| clade_name            |             |
+| taxonomy_id           |             |
+| relative_abundance \* |             |
+| additional_species    |             |
 
-> * Value used in standardised profile output, multiplied with a fixed number to yield an integer for the standardised profile
+> \* Value used in standardised profile output, multiplied with a fixed number to yield an integer for the standardised profile
 
 ## Example
 
@@ -27,4 +27,3 @@ k__Bacteria|p__Bacteroidetes	2|976	65.30411
 k__Bacteria|p__Proteobacteria	2|1224	27.81459
 k__Bacteria|p__Firmicutes	2|1239	4.35836
 ```
-
