+++ docs/tutorial.md
@@ -1,8 +1,12 @@
 # Tutorial
 
-<!-- TODO background -->
-
 In this tutorial we will show you how to generate taxonomic profiles using two popular taxonomic classifers, `Kraken2` and `MetaPhlAn3`, demonstrate how their profile output is very different, and how to make them more easily comparable using `taxpasta`.
+
+The motivation behind `taxpasta` is the problem of ... <!-- TODO background -->
+
+Each profiler has their own advantages and disadvantages, just because one profiler picks up more taxa, does not necessarily mean they are more accurate!
+
+However it is also important to note that different profilers produce additional output columns that may allow you to make more informed decision to the reliability or confidence of each taxonomic identification. So while `taxpasta` is a utility to make comparison between profilers easier to be performed, this does not necessarily mean all comparisons are necessarily valid.
 
 ## Preparation
 
@@ -58,6 +62,8 @@ With the software, database, and reads all successfully downloaded and installed
 
 ## Generate taxonomic profiles
 
+<!-- save to profiles/ -->
+
 ## Taxonomic profile output comparison
 
 ## Standardising with taxpasta
@@ -78,6 +84,7 @@ Finally we supply a list of output files from each of the classifers.
 As a whole, the commands for both tools should look like the following
 
 ```bash
+cd profiles
 taxpasta merge --profiler kraken2 --output kraken2_taxon_table.tsv sample1.kreport.txt sample2.kreport.txt sample3.kreport.txt
 taxpasta merge --profiler metaphlan --output metaphlan3_taxon_table.tsv sample1.mp3.txt sample2.mp3.txt sample3.mp3.txt
 ```
@@ -88,21 +95,86 @@ You can then view the output by simply running `head` on the two output files
 head kraken2_taxon_table.tsv
 ```
 
-```bash
-
+```text
+<!-- TODO -->
 ```
 
 ```bash
-head metaphlan_taxon_table.tsv
+head kraken2_taxon_table.tsv
+```
+
+```text
+<!-- TODO -->
 ```
 
 > ℹ️ If you wish to produce only per-sample standarised tables, and merge them into a single table yourslf later, use `taxpasta standardise`.
 
-We can then more easily load
+By default, `taxpasta` will simply use the file name as the 'sample name' ID in the headers of each column. If you wish to specify your own sample name for each file, you can specify to `taxpasta` a 'samplesheet' that contains both the paths to the profiles and the sample name. To demonstrate this, we can use a little bit of command-land trickier to generate such a samplesheet.
 
-> ⚠️ Each profiler has their own advantages and disadvantages, just because one profiler picks up more taxa, does not necessarily mean they are more accurate! Furthermore, different profilers produce additional output columns that may allow you to make more informed decision to the reliability or confidence of each taxonomic identification.
+```bash
+cd profiles/
 
-> ⚠️ You must make sure that the _taxonomy ID_ used in the databases the two classfiers are the same! I.e., the taxonomic IDs refer to the same taxon reference sequences!
+## Get the full paths for each file
+ls -1 *kreport.txt > kraken2_paths.txt
+
+## Construct a sample name based on the filename
+sed 's#kreport.txt##g' paths.txt > kraken2_names.txt
+
+## Create the samplesheet, adding a header and then adding the samplenames and paths
+echo 'sample profile' > kraken2_samplesheet.txt
+paste -d " " kraken2_names.txt kraken2_paths.txt >> kraken2_samplesheet.txt
+```
+
+To check the file has been created correctly
+
+```bash
+cat kraken2_samplesheet.txt
+```
+
+We can then re-run `taxpasta` but with the samplesheet rather than the list of files at the end
+
+```bash
[…]
