### 2023-03-08 #84 @jfy133
> Also @sofstam @jfy133 could you both please fill out the pre-review survey? I have a reply for @Midnighter but not you I think. https://forms.gle/F9mou7S3jhe8DMJ16 It's just a brief (5-10m) survey to help us understand how we're doing as an org. Thank you! pray

oops, sorry - from the second page most of the questions seemed to be about the package and I assume Moritz had already provided all that info.

I've filled it out, but left pretty much all the optional questions empty as either you have that info from Moritz, and/or submitting via pyOpenSci is Moritz' initative so I'm not that familiar/involved with python etc.

One comment though: I noticed that the options in your 'What background best describes you cultural identity?' question is extremely N. American focused. Unfortunately I don't have a good solution for you, but you may not be getting a particularly good overview of this - e.g. Asian spans half the world with many equally large sub-divisons as the Pacific Islanders, and also grouping 'Black' and African-American is also arguably unfair as they also can have a large difference in backgrounds/problems etc. For example see what official surveys from UK use: https://www.ethnicity-facts-figures.service.gov.uk/style-guide/ethnic-groups
