### 2023-03-08 1be9e90
docs: expand contributing also to full documentation
