### 2023-03-08 8fad161
docs: shorten front matter
