### 2023-03-08 1adcdd1
docs: expand the installation section
