### 2023-03-08 #84 @Midnighter
@NickleDave regarding your comments on `taxpasta -h` the truncation depends on your terminal width and is automatically created by typer/rich. I'll look into shortening the summaries such that they fit within smaller windows without truncation.
