### 2023-03-08 08ae56b
docs: use mkdocs-material footnotes
