+++ docs/supported_profilers/metaphlan.md
@@ -6,14 +6,14 @@
 
 Taxpasta expects a four-column, tab-separated file. This is generated by redirecting the MetaPhlAn3 `stdout`. It interprets the columns as:
 
-| Column Header         | Description |
-| --------------------- | ----------- |
-| clade_name            |             |
-| taxonomy_id           |             |
-| relative_abundance \* |             |
-| additional_species    |             |
+| Column Header          | Description |
+| ---------------------- | ----------- |
+| clade_name             |             |
+| taxonomy_id            |             |
+| relative_abundance[^1] |             |
+| additional_species     |             |
 
-> \* Value used in standardised profile output, multiplied with a fixed number to yield an integer for the standardised profile
+[^1]: Value used in standardised profile output, multiplied with a fixed number to yield an integer for the standardised profile
 
 ## Example
 
