+++ docs/supported_profilers/kaiju.md
@@ -10,11 +10,11 @@ Taxpasta expects a five column output. This is generated by the `kaiju2table` co
 | ------------- | ----------- |
 | file          |             |
 | percent       |             |
-| reads \*      |             |
+| reads[^1]     |             |
 | taxon_id      |             |
 | taxon_name    |             |
 
-> \* Value used in standardised profile output
+[^1]: Value used in standardised profile output
 
 ## Example
 
