### 2023-03-08 199f93b
docs: use mkdocs-material footnotes
