+++ docs/supported_profilers/megan6.md
@@ -15,11 +15,11 @@ Taxpasta expects a two-column, tab-separated file. This is generated either by r
 | Column Header | Description |
 | ------------- | ----------- |
 | taxonomy_id   |             |
-| count \*      |             |
+| count[^1]     |             |
 
 Note that the file itself does not have any column names.
 
-> \* Value used in standardised profile output
+[^1]: Value used in standardised profile output
 
 ## Example
 
