### 2023-03-08 #84 @Midnighter
> Would it be worth making rich a dependency so you don't have to special case typer behavior?

At the moment, it's an extra dependency that can be installed with `taxpasta[rich]`. (Come to think of it, we should document all of those.) The reason that it's not default is that where taxpasta originated, as a CLI tool for a nextflow pipeline, fancy terminal output is really not needed. We should definitely recommend installing that for interactive work, though.
