### 2023-03-08 80151bf
fix: provide files directly to `TaxDb`
Providing only a directory, prompts taxopy to delete its contents and
download fresh dump files from NCBI. This is not the intended behavior.
We want to use local files and keep them.
