### 2023-03-08 #84 @NickleDave
> where taxpasta originated, as a CLI tool for a nextflow pipeline, fancy terminal output is really not needed.

Understood!

Btw @Midnighter could I ask you to link this review issue on any issues you raise to make changes before review? By adding a link to the issue on the taxpasta repo, so that GitHub cross references them.

Just to help us track. Thank you :pray:
