### 2023-03-08 8badddc
docs: prepare changelog for 0.2.2 release
