+++ docs/adr/0002-use-integer-taxonomy-identifiers.md
@@ -27,4 +27,3 @@ later.
 
 Another consequence is that if we ever want to attach names or lineages to our
 output files, the corresponding taxonomy will have to be loaded separately.
-
