### 2023-03-08 taxprofiler/taxpasta#68 @Midnighter
Add the Zenodo webhook and badge
Based on comment from https://github.com/pyOpenSci/software-submission/issues/84
