### 2023-03-08 taxprofiler/taxpasta#68 @Midnighter
Fixed in 7c84b3aa48ea5928a5cd6287327c05d456e11afa
