+++ README.md
@@ -47,7 +47,7 @@ and thus automatically generated
 
 ## Copyright
 
--   Copyright © 2022, Moritz E. Beber, Maxime Borry, James A. Fellows Yates, and
+-   Copyright © 2022, 2023, Moritz E. Beber, Maxime Borry, James A. Fellows Yates, and
     Sofia Stamouli.
 -   Free software distributed under the [Apache Software License
     2.0](https://www.apache.org/licenses/LICENSE-2.0).
