### 2023-03-08 taxprofiler/taxpasta#69 @Midnighter
Document extras in installation docs
Based on comment from https://github.com/pyOpenSci/software-submission/issues/84
