### 2023-03-08 14ef7d4
docs: describe optional installation dependencies
