### 2023-03-08 df7b8eb
docs: describe optional installation dependencies
