### 2023-03-08 034ef91
chore: rename 'development' dependencies to 'dev'
