+++ README.md
@@ -61,7 +61,7 @@ pip install 'taxpasta[rich,biom]'
 -   `ods` supports writing output tables in [ODS](https://www.libreoffice.org/discover/what-is-opendocument/) format.
 -   `xlsx` supports writing output tables in [Microsoft Excel](https://support.microsoft.com/en-us/office/file-formats-that-are-supported-in-excel-0943ff2c-6014-4e8d-aaea-b83d51d46247) format.
 -   `all` includes all of the above.
--   `development` provides all tools needed for contributing to taxpasta.
+-   `dev` provides all tools needed for contributing to taxpasta.
 
 ## Copyright
 
