+++ docs/tutorial.qmd
@@ -1,114 +1,0 @@
----
-title: "Tutorial"
-author: "James A. Fellows Yates"
-format: html
----
-
-## Introduction
-
-This tutorial will show you how to generate standardised taxonomic profiles from the heterogeneous output of two popular taxonomic profilers: [`Kraken2`](https://ccb.jhu.edu/software/kraken2/) and [`mOTUs`](https://motu-tool.org/), and then demonstrate some of the benefits of this standardisation when running downstream analyses on such tables in the popular data science programming language [`R`](https://www.r-project.org/).
-
-## Preparation
-
-For this tutorial you will need an internet connection, `taxpasta` already [installed](/index), and  R with the `readr` package from the [Tidyverse](https://tidyverse.org) set of packages installed, and a UNIX based operating system (Linux, OSX or Windows Subsystem for Linux).
-
-First we will make a 'scratch' directory where we can run all the tutorial and delete the files after.
-
-```{bash}
-mkdir taxpasta-tutorial
-```
-
-We will also need to download some example taxonomic profiles from `Kraken2` and `mOTUs`. If you're on UNIX a machine we can download some test data from the taxpasta repository.
-
-```{bash}
-## mOTUs
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out
-
-## Kraken2
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
-```
-
-## Tutorial
-
-To begin, we can try loading a `mOTUs` profile into R using a common and 'default' table reading command, `read_tsv()` from the `readr` package.
-
-```{r}
-library(readr)
-
-profile_motus_2612_pe <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out")
-```
-
-You can see we immediately hit an error, as there is a 'comment' line at the top of the profile with information on how the profile was generated.
-
-While this is very nice for reproducibility, to load this we have to instead add extra options to the function, which makes loading the table less than smooth for downstream analyses.
-
-```{r}
-profile_motus_2612_pe <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out", comment = "#")
-```
-
-However, once again we hit another error - the column headers are _also_ specified as a comment line...
-
-Instead we can try to skip the first two tool information lines
-
-```{r}
-profile_motus_2612_pe <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out", skip = 2)
-
-profile_motus_2612_pe 
-```
-
-This now works! However we argue getting this to work takes much too effort to simply load what is essentially a simple table.
-
-Furthermore, we would have to load each profile one by one for each sample, requiring more complicated loops and table join code.
-
-Now lets try loading the `Kraken2` output.
-
-```{r}
-profile_kraken2_2612_pe <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt")
-profile_kraken2_2612_pe
-```
- 
[…]
