+++ docs/tutorial.qmd
@@ -10,33 +10,55 @@ This tutorial will show you how to generate standardised taxonomic profiles from
 
 ## Preparation
 
-For this tutorial you will need an internet connection, `taxpasta` already [installed](/index), and  R with the `readr` package from the [Tidyverse](https://tidyverse.org) set of packages installed, and a UNIX based operating system (Linux, OSX or Windows Subsystem for Linux).
+### Software
+
+For this tutorial you will need an internet connection, `taxpasta` already [installed](/index), and  R with the `readr` and `dplyr` packages from the [Tidyverse](https://tidyverse.org) set of packages installed, and a UNIX based operating system (Linux, OSX or Windows Subsystem for Linux).
+
+To summarise you will need:
+
+- Unix terminal (e.g. `bash`)
+- `taxpasta`
+- `R`
+  - `readr` package
+  - `dplyr` package
+
+### Data
 
 First we will make a 'scratch' directory where we can run all the tutorial and delete the files after.
 
-```{bash}
+```{bash, eval=F}
 mkdir taxpasta-tutorial
 ```
 
+
 We will also need to download some example taxonomic profiles from `Kraken2` and `mOTUs`. If you're on UNIX a machine we can download some test data from the taxpasta repository.
 
-```{bash}
+```{bash, eval=F}
+which conda
 ## mOTUs
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out 
+curl -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
 
 ## Kraken2
-curl https://github.com/taxprofiler/taxpasta/raw/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt 
+```
+
+We should now see three files with contents in the `taxpasta-tutorial` directory
+
+```{bash, eval=F}
+ls -l taxpasta-tutorial/*
 ```
 
 ## Tutorial
+
+### Raw classifer output
 
 To begin, we can try loading a `mOTUs` profile into R using a common and 'default' table reading command, `read_tsv()` from the `readr` package.
 
 ```{r}
 library(readr)
 
-profile_motus_2612_pe <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out")
+profile_motus_2612_pe_raw <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out")
 ```
 
 You can see we immediately hit an error, as there is a 'comment' line at the top of the profile with information on how the profile was generated.
@@ -44,7 +66,7 @@ You can see we immediately hit an error, as there is a 'comment' line at the top
[…]
