### 2023-03-09 87c3f2d
chore: add docs requirements as extra
