+++ docs/requirements.txt
@@ -1,3 +1,0 @@
-mkdocs-material ~=9.1
-mkdocstrings[python] ~=0.20
-mkdocs-awesome-pages-plugin ~=2.8
