### 2023-03-09 fc3bc14
chore: add docs requirements as extra
