+++ pyproject.toml
@@ -73,7 +73,12 @@ rich = ["rich"]
 xlsx = ["openpyxl"]
 dev = [
     "pre-commit",
-    "tox ~=4.0"
+    "tox ~=4.0",
+]
+docs = [
+    "mkdocs-material ~=9.1",
+    "mkdocstrings[python] ~=0.20",
+    "mkdocs-awesome-pages-plugin ~=2.8",
 ]
 
 ################################################################################
