+++ .readthedocs.yaml
@@ -12,4 +12,7 @@ mkdocs:
 
 python:
    install:
-   - requirements: docs/requirements.txt
+   - method: pip
+     path: "."
+     extra_requirements:
+       - docs
