### 2023-03-09 4bad8cf
docs: document the new 'docs' extra
