### 2023-03-09 8e09071
docs: document the new 'docs' extra
