### 2023-03-09 8af432f
style: format readthedocs config
