+++ .readthedocs.yaml
@@ -11,8 +11,8 @@ mkdocs:
   configuration: mkdocs.yml
 
 python:
-   install:
-   - method: pip
-     path: "."
-     extra_requirements:
-       - docs
+  install:
+    - method: pip
+      path: "."
+      extra_requirements:
+        - docs
