### 2023-03-09 e9f2aff
fix: let tox install from docs extras
