+++ tox.ini
@@ -53,9 +53,7 @@ commands=
     mypy {toxinidir}/src/taxpasta
 
 [testenv:docs]
-skip_install = True
-deps=
-    -r{toxinidir}/docs/requirements.txt
+extras = docs
 commands=
     mkdocs build --strict
 
