+++ .readthedocs.yaml
@@ -15,4 +15,4 @@ python:
     - method: pip
       path: "."
       extra_requirements:
-        - docs
+        - dev
