+++ tox.ini
@@ -53,7 +53,7 @@ commands=
     mypy {toxinidir}/src/taxpasta
 
 [testenv:docs]
-extras = docs
+extras = dev
 commands=
     mkdocs build --strict
 
