+++ docs/tutorial.md
@@ -1,0 +1,546 @@
+Tutorial
+================
+James A. Fellows Yates
+
+## Introduction
+
+This tutorial will show you how to generate standardised taxonomic
+profiles from the heterogeneous output of two popular taxonomic
+profilers: [`Kraken2`](https://ccb.jhu.edu/software/kraken2/) and
+[`mOTUs`](https://motu-tool.org/), and then demonstrate some of the
+benefits of this standardisation when running downstream analyses on
+such tables in the popular data science programming language
+[`R`](https://www.r-project.org/).
+
+## Preparation
+
+### Software
+
+For this tutorial you will need an internet connection, `taxpasta`
+already [installed](/index), and R with the `readr` and `dplyr` packages
+from the [Tidyverse](https://tidyverse.org) set of packages installed,
+and a UNIX based operating system (Linux, OSX or Windows Subsystem for
+Linux).
+
+To summarise you will need:
+
+-   Unix terminal (e.g. `bash`)
+-   `taxpasta`
+-   `R`
+    -   `readr` package
+    -   `dplyr` package
+
+### Data
+
+First we will make a ‘scratch’ directory where we can run all the
+tutorial and delete the files after.
+
+``` bash
+mkdir taxpasta-tutorial
+```
+
+We will also need to download some example taxonomic profiles from
+`Kraken2` and `mOTUs`. If you’re on UNIX a machine we can download some
+test data from the taxpasta repository.
+
+``` bash
+## mOTUs
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out 
+curl -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
+
+## Kraken2
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt 
+```
+
+We should now see three files with contents in the `taxpasta-tutorial`
+directory
+
+``` bash
+ls -l taxpasta-tutorial/*
+```
+
+## Tutorial
+
+### Raw classifer output
+
+To begin, we can try loading a `mOTUs` profile into R using a common and
+‘default’ table reading command, `read_tsv()` from the `readr` package.
+
+``` r
+library(readr)
+
+profile_motus_2612_pe_raw <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out")
+```
+
+    Warning: One or more parsing issues, see `problems()` for details
+
+    Rows: 33573 Columns: 1
+    ── Column specification ────────────────────────────────────────────────────────
+    Delimiter: "\t"
+    chr (1): # git tag version 3.0.3 |  motus version 3.0.3 | map_tax 3.0.3 | ge...
+
+    ℹ Use `spec()` to retrieve the full column specification for this data.
+    ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
+
+You can see we immediately hit an error, as there is a ‘comment’ line at
+the top of the profile with information on how the profile was
+generated.
+
+While this is very nice for reproducibility, to load this we have to
+instead add extra options to the function, which makes loading the table
+less than smooth for downstream analyses.
+
+``` r
[…]
