### 2023-03-09 7d0b64e
Add bash block rendering
