+++ docs/tutorial.qmd
@@ -26,25 +26,25 @@ To summarise you will need:
 
 First we will make a 'scratch' directory where we can run all the tutorial and delete the files after.
 
-```{bash, eval=F}
+```{bash}
 mkdir taxpasta-tutorial
 ```
 
 
 We will also need to download some example taxonomic profiles from `Kraken2` and `mOTUs`. If you're on UNIX a machine we can download some test data from the taxpasta repository.
 
-```{bash, eval=F}
+```{bash}
 ## mOTUs
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out 
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
 curl -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
 
 ## Kraken2
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt 
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
 ```
 
 We should now see three files with contents in the `taxpasta-tutorial` directory
 
-```{bash, eval=F}
+```{bash}
 ls -l taxpasta-tutorial/*
 ```
 
@@ -52,7 +52,23 @@ ls -l taxpasta-tutorial/*
 
 ### Raw classifer output
 
-To begin, we can try loading a `mOTUs` profile into R using a common and 'default' table reading command, `read_tsv()` from the `readr` package.
+To begin, lets look at the contents of the output from each of the profilers
+
+For `mOTUs`:
+
+```bash
+head taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
+```
+
+and `Kraken2`:
+
+```bash
+head taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
+```
+
+These look quite different, and neither are in a nice 'pure' tabular formats that data scientists and analysis software normally likes. They also have different types columns, and in the case of `Kraken2` has an interesting 'indentation' way of showing the taxonomic rank of each hit.
+
+We can try loading a `mOTUs` profile into R using a common and 'default' table reading command, `read_tsv()` from the `readr` package.
 
 ```{r}
 library(readr)
@@ -88,7 +104,7 @@ Now lets try loading the `Kraken2` output.
 profile_kraken2_2612_pe_raw <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt")
 profile_kraken2_2612_pe_raw
 ```
- 
+
 This doesn't fail to load, but unfortunately the column headers look a bit wierd... it seems the Kraken2 'table' file does not include a column header! In this case we have to specify these ourselves..
 
 ```{r}
@@ -126,11 +142,19 @@ But this is where `taxpasta` comes to the rescue!
 
 With `taxpasta`, we can already standardised and make multi-sample taxon tables for you at the command line level (e.g., immediately after profiling), rather than having to do this with custom scripts and a lot of manual munging.
 
[…]
