### 2023-03-09 114a6e1
Add bash block rendering
