+++ docs/tutorial.md
@@ -45,12 +45,30 @@ test data from the taxpasta repository.
 
 ``` bash
 ## mOTUs
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out 
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
 curl -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
 
 ## Kraken2
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt 
+curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
 ```
+
+      % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
+                                     Dload  Upload   Total   Spent    Left  Speed
+
+      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
+      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
+    100 1967k  100 1967k    0     0  3218k      0 --:--:-- --:--:-- --:--:-- 3215k
+      % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
+                                     Dload  Upload   Total   Spent    Left  Speed
+
+      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
+      1 1967k    1 25578    0     0  65025      0  0:00:30 --:--:--  0:00:30 64918
+    100 1967k  100 1967k    0     0  4039k      0 --:--:-- --:--:-- --:--:-- 4032k
+      % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
+                                     Dload  Upload   Total   Spent    Left  Speed
+
+      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
+    100  2681  100  2681    0     0   6250      0 --:--:-- --:--:-- --:--:--  6264
 
 We should now see three files with contents in the `taxpasta-tutorial`
 directory
@@ -59,12 +77,36 @@ directory
 ls -l taxpasta-tutorial/*
 ```
 
+    -rw-rw-r-- 1 jfellows jfellows    2681 Mar  9 15:25 taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
+    -rw-rw-r-- 1 jfellows jfellows 2015168 Mar  9 15:25 taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
+    -rw-rw-r-- 1 jfellows jfellows 2015125 Mar  9 15:25 taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out
+
 ## Tutorial
 
 ### Raw classifer output
 
-To begin, we can try loading a `mOTUs` profile into R using a common and
-‘default’ table reading command, `read_tsv()` from the `readr` package.
[…]
