### 2023-03-10 a6ae6f2
Apply suggestions from code review
Co-authored-by: Sofia Stamouli <91951607+sofstam@users.noreply.github.com>
