+++ docs/tutorials/index.md
@@ -1,0 +1,13 @@
+# Tutorials
+
+The following pages provide longer-more step-by-step tutorials on how to run and use the various functionality of `taxpasta`.
+
+## General tutorials
+
+-   [Introductory Tutorial](tutorial.md)
+    -   How to start from raw outputs and generate standardised and merged profiles
+
+## How Tos
+
+-   [How To: Adding Taxon Names To Standardised Output](howto-adding-taxon-names.md)
+-   [How To: Using Custom Sample Names](howto-custom-sample-names.md)
