+++ docs/tutorials/tutorial.md
@@ -1,12 +1,12 @@
-Tutorial
-================
-James A. Fellows Yates
+
+# General Tutorial
 
 ## Introduction
 
 This tutorial will show you how to generate standardised taxonomic
 profiles from the heterogeneous output of two popular taxonomic
-profilers: [`Kraken2`](https://ccb.jhu.edu/software/kraken2/) and
+classifiers/profilers:
+[`Kraken2`](https://ccb.jhu.edu/software/kraken2/) and
 [`mOTUs`](https://motu-tool.org/), and then demonstrate some of the
 benefits of this standardisation when running downstream analyses on
 such tables in the popular data science programming language
@@ -24,11 +24,11 @@ Linux).
 
 To summarise you will need:
 
--   Unix terminal (e.g. `bash`)
--   `taxpasta`
--   `R`
-    -   `readr` package
-    -   `dplyr` package
+- Unix terminal (e.g. `bash`)
+- `taxpasta`
+- `R`
+  - `readr` package
+  - `dplyr` package
 
 ### Data
 
@@ -56,19 +56,18 @@ curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.
                                      Dload  Upload   Total   Spent    Left  Speed
 
       0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-    100 1967k  100 1967k    0     0  3218k      0 --:--:-- --:--:-- --:--:-- 3215k
+     41 1967k   41  824k    0     0  1940k      0  0:00:01 --:--:--  0:00:01 1940k
+    100 1967k  100 1967k    0     0  3566k      0 --:--:-- --:--:-- --:--:-- 3565k
       % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                      Dload  Upload   Total   Spent    Left  Speed
 
       0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-      1 1967k    1 25578    0     0  65025      0  0:00:30 --:--:--  0:00:30 64918
-    100 1967k  100 1967k    0     0  4039k      0 --:--:-- --:--:-- --:--:-- 4032k
+    100 1967k  100 1967k    0     0  5448k      0 --:--:-- --:--:-- --:--:-- 5451k
       % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                      Dload  Upload   Total   Spent    Left  Speed
 
       0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-    100  2681  100  2681    0     0   6250      0 --:--:-- --:--:-- --:--:--  6264
+    100  2681  100  2681    0     0  20671      0 --:--:-- --:--:-- --:--:-- 20782
 
 We should now see three files with contents in the `taxpasta-tutorial`
 directory
@@ -77,16 +76,16 @@ directory
 ls -l taxpasta-tutorial/*
 ```
 
-    -rw-rw-r-- 1 jfellows jfellows    2681 Mar  9 15:25 taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
-    -rw-rw-r-- 1 jfellows jfellows 2015168 Mar  9 15:25 taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
[…]
