### 2023-03-12 ce9269c
fix: correctly set the version according to hatch-vcs
