+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-  Fixed the version string according to hatch-vcs output (#74).
+
 ## [0.2.2] - (2023-03-08)
 
 ### Fixed
