### 2023-03-12 cc71d2e
tests: add cases for the version string
