### 2023-03-12 52a596e
tests: add cases for the version string
