+++ tests/conftest.py
@@ -22,6 +22,13 @@
 from pathlib import Path
 
 import pytest
+from typer.testing import CliRunner
+
+
+@pytest.fixture(scope="session")
+def runner() -> CliRunner:
+    """Return a CLI runner instance for testing."""
+    return CliRunner(mix_stderr=False)
 
 
 @pytest.fixture(scope="session")
