### 2023-03-12 2c9fae9
fix: increase the tolerance when comparing floats
