### 2023-03-12 695796f
tests: adjust failure cases to new default tolerance
