### 2023-03-12 d2d3b48
docs: prepare changelog for release 0.2.3
