### 2023-03-13 c177970
refactor: include Python, R, and bash snippets
Include snippets from markdown files, that were previously rendered from
Quarto markdown, in the main tutorial file.
