+++ docs/tutorials/tutorial.qmd
@@ -1,228 +1,0 @@
----
-format: gfm
----
-
-# General Tutorial
-
-## Introduction
-
-This tutorial will show you how to generate standardised taxonomic profiles from the heterogeneous output of two popular taxonomic classifiers/profilers: [`Kraken2`](https://ccb.jhu.edu/software/kraken2/) and [`mOTUs`](https://motu-tool.org/), and then demonstrate some of the benefits of this standardisation when running downstream analyses on such tables in the popular data science programming language [`R`](https://www.r-project.org/).
-
-## Preparation
-
-### Software
-
-For this tutorial you will need an internet connection, `taxpasta` already [installed](/index), and  R with the `readr` and `dplyr` packages from the [Tidyverse](https://tidyverse.org) set of packages installed, and a UNIX based operating system (Linux, OSX or Windows Subsystem for Linux).
-
-To summarise you will need:
-
-- Unix terminal (e.g. `bash`)
-- `taxpasta`
-- `R`
-  - `readr` package
-  - `dplyr` package
-
-### Data
-
-First we will make a 'scratch' directory where we can run all the tutorial and delete the files after.
-
-```{bash}
-mkdir taxpasta-tutorial
-```
-
-
-We will also need to download some example taxonomic profiles from `Kraken2` and `mOTUs`. If you're on UNIX a machine we can download some test data from the taxpasta repository.
-
-```{bash}
-## mOTUs
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
-curl -o taxpasta-tutorial/2612_se-ERR5766180-db_mOTU.out https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
-
-## Kraken2
-curl -o taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
-```
-
-We should now see three files with contents in the `taxpasta-tutorial` directory
-
-```{bash}
-ls -l taxpasta-tutorial/*
-```
-
-## Tutorial
-
-### Raw classifer output
-
-To begin, let's look at the contents of the output from each of the classifiers/profilers.
-
-For `mOTUs`:
-
-```bash
-head taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out
-```
-
-and `Kraken2`:
-
-```bash
-head taxpasta-tutorial/2612_pe-ERR5766176-db1.kraken2.report.txt
-```
-
-These look quite different, and neither are in a nice "pure" tabular formats that data scientists and analysis software normally like. They also have different types columns, and in the case of `Kraken2` has an interesting "indentation" way of showing the taxonomic rank of each hit.
-
-We can try loading a `mOTUs` profile into R using a common and "default" table reading command, `read_tsv()` from the `readr` package.
-
-```{r}
-library(readr)
-
-profile_motus_2612_pe_raw <- read_tsv("taxpasta-tutorial/2612_pe-ERR5766176-db_mOTU.out")
-```
-
-You can see we immediately hit an error, as there is a 'comment' line at the top of the profile with information on how the profile was generated.
-
-While this is very nice for reproducibility, to load this we have to instead add extra options to the function, which makes loading the table less than smooth for downstream analyses.
-
-```{r}
[…]
