+++ docs/tutorials/getting-started.md
@@ -1,0 +1,273 @@
+# Getting Started
+
+## Introduction
+
+We will show you how to generate standardised taxonomic profiles from the
+diverse outputs of two popular taxonomic profilers:
+[Kraken2](https://ccb.jhu.edu/software/kraken2/) and
+[mOTUs](https://motu-tool.org/). We demonstrate some benefits of this
+standardisation when running downstream analyses on such tables in either the
+popular statistical programming language [R](https://www.r-project.org/) or
+[Python](https://www.python.org/).
+
+## Preparation
+
+### Software
+
+For this tutorial you will need an internet connection, an [installation of
+taxpasta](/#install), and, if you want to follow the R parts, an [installation
+of
+R](https://rstudio-education.github.io/hopr/starting.html#how-to-download-and-install-r)
+itself with the [readr](https://readr.tidyverse.org/) and
+[dplyr](https://dplyr.tidyverse.org/) packages from the
+[Tidyverse](https://tidyverse.org). Furthermore, we assume a UNIX based
+operating system, such as Linux, macOS, or Windows Subsystem for Linux.
+
+To summarise, you will need:
+
+
+=== "R"
+
+--8<--
+    tutorial_r_snippets.md:software
+--8<--
+
+=== "Python"
+
+--8<--
+    tutorial_python_snippets.md:software
+--8<--
+
+### Data
+
+First we will make a ‘scratch’ directory where we can run the tutorial and
+delete again afterwards.
+
+--8<--
+tutorial_bash_snippets.md:data-setup
+--8<--
+
+We will also need to download some example taxonomic profiles from Kraken2 and
+mOTUs. We can download test data from the taxpasta repository using, for
+example, `curl` or `wget`.
+
+=== "curl"
+
+--8<--
+    tutorial_bash_snippets.md:data-curl
+--8<--
+
+=== "wget"
+
+--8<--
+    tutorial_bash_snippets.md:data-wget
+--8<--
+
+We should now see three files with contents in the `taxpasta-tutorial` directory
+
+--8<--
+tutorial_bash_snippets.md:data-ls
+--8<--
+
+## Profiles
+
+### Raw Output
+
+To begin, let’s look at the contents of the output from each profiler.
+
+=== "mOTUs"
+
+--8<--
+    tutorial_bash_snippets.md:motus-head
+--8<--
+
+=== "Kraken2"
+
+--8<--
+    tutorial_bash_snippets.md:kraken2-head
+--8<--
+
+These look quite different and neither of them is in a nice "pure" tabular
+format that is convenient for analysis software to load. They also have
+different types columns and, in the case of Kraken2, has an interesting
+"indentation" way of showing the taxonomic rank of each taxon.
+
+#### mOTUs
+
+=== "R"
+
+--8<--
+    tutorial_r_snippets.md:raw-motus
+--8<--
+
+=== "Python"
+
+--8<--
+    tutorial_python_snippets.md:raw-motus
+--8<--
+
+That works! At least in terms of reasonable table headers. However, we can see
+that there are missing NCBI taxonomy identifiers so there is probably more data
+cleaning ahead of us.  Getting to this point took too much effort already to
+load what is essentially a simple table. Furthermore, if we are interested in
+loading all mOTUs profiles, we would have to load each profile one by one for
+each sample, requiring more complicated loops and table join code.
+
+#### Kraken2
+
+=== "R"
+
+--8<--
+    tutorial_r_snippets.md:raw-kraken2
+--8<--
+
+=== "Python"
+
+--8<--
+    tutorial_python_snippets.md:raw-kraken2
+--8<--
+
+This looks better but if we look at the [Kraken2
[…]
