### 2023-03-13 11673dd
chore: create docs dependencies in tests
