### 2023-03-13 ec1d862
chore: create docs dependencies in tests
