### 2023-03-13 72f5573
chore: create help output on the fly
