### 2023-03-13 d5136a7
chore: create help output on the fly
