+++ tox.ini
@@ -6,7 +6,7 @@ python =
     3.8: install, safety, py38
     3.9: install, safety, py39
     3.10: install, safety, py310
-    3.11: bandit, flake8, install, safety, mpypy, docs, py311
+    3.11: bandit, flake8, install, safety, mpypy, py311
 
 [testenv]
 # Define base for all test environments.
