### 2023-03-13 f31808d
chore: manually generate `make docs`
