+++ docs/quick_reference/consensus_help.txt
@@ -1,0 +1,9 @@
+                                                                                              
+ Usage: taxpasta consensus [OPTIONS]                                                          
+                                                                                              
+ Form a consensus for the same sample but from different taxonomic profiles.                  
+                                                                                              
+╭─ Options ──────────────────────────────────────────────────────────────────────────────────╮
+│ --help  -h        Show this message and exit.                                              │
+╰────────────────────────────────────────────────────────────────────────────────────────────╯
+
