+++ docs/tutorials/tutorial_python_snippets.md
@@ -1,0 +1,195 @@
+<!-- --8<-- [start:software] -->
+
+-   Unix terminal (e.g., `bash`)
+-   [Python environment with taxpasta](/#install)
+-   No additional requirements, since we will make use of
+    [pandas](https://pandas.pydata.org/) which is already part of the
+    taxpasta installation.
+
+<!-- --8<-- [end:software] -->
+<!-- --8<-- [start:raw-motus] -->
+
+We can try loading a mOTUs profile into Python using the common table
+reading function `read_table()` from the `pandas` package with default
+arguments.
+
+```python
+import pandas as pd
+
+profile_motus_2612_pe_raw = pd.read_table("2612_pe-ERR5766176-db_mOTU.out")
+```
+
+    ParserError: Error tokenizing data. C error: Expected 1 fields in line 3, saw 3
+
+You can see we immediately hit a `ParserError`, as there is a ‘comment’
+line at the top of the mOTUs profile with information on how the profile
+was generated.
+
+While such a comment is very useful for reproducibility, to load this we
+have to instead add extra options to the function, which makes loading
+the table less than smooth for downstream analyses.
+
+```python
+profile_motus = pd.read_table(
+    "2612_pe-ERR5766176-db_mOTU.out",
+    comment="#",
+)
+profile_motus.head()
+```
+
+|     | Leptospira alexanderi \[ref_mOTU_v3_00001\]                  |      100053 |   0 |
+| --: | :----------------------------------------------------------- | ----------: | --: |
+|   0 | Leptospira weilii \[ref_mOTU_v3_00002\]                      |       28184 |   0 |
+|   1 | Chryseobacterium sp. \[ref_mOTU_v3_00004\]                   |         nan |   0 |
+|   2 | Chryseobacterium gallinarum \[ref_mOTU_v3_00005\]            | 1.32435e+06 |   0 |
+|   3 | Chryseobacterium indologenes \[ref_mOTU_v3_00006\]           |         253 |   0 |
+|   4 | Chryseobacterium artocarpi/ureilyticum \[ref_mOTU_v3_00007\] |         nan |   0 |
+
+However, once again we encounter another problem: the column headers are
+_also_ specified as a comment line and our first line of data is instead
+read as the column headers. We can try to skip the first two lines
+entirely to remedy this.
+
+```python
+profile_motus = pd.read_table(
+    "2612_pe-ERR5766176-db_mOTU.out",
+    skiprows=2,
+)
+profile_motus.head()
+```
+
+|     | \#consensus_taxonomy                               | NCBI_tax_id | 2612_pe-ERR5766176-db_mOTU |
+| --: | :------------------------------------------------- | ----------: | -------------------------: |
+|   0 | Leptospira alexanderi \[ref_mOTU_v3_00001\]        |      100053 |                          0 |
+|   1 | Leptospira weilii \[ref_mOTU_v3_00002\]            |       28184 |                          0 |
+|   2 | Chryseobacterium sp. \[ref_mOTU_v3_00004\]         |         nan |                          0 |
+|   3 | Chryseobacterium gallinarum \[ref_mOTU_v3_00005\]  | 1.32435e+06 |                          0 |
+|   4 | Chryseobacterium indologenes \[ref_mOTU_v3_00006\] |         253 |                          0 |
+
+<!-- --8<-- [end:raw-motus] -->
+<!-- --8<-- [start:raw-kraken2] -->
+
+```python
+profile_kraken2 = pd.read_table("2612_pe-ERR5766176-db1.kraken2.report.txt")
+profile_kraken2.head()
+```
+
+|     | 99.97 | 627680 | 627680.1 | U   |      0 | unclassified       |
[…]
