+++ docs/tutorials/tutorial_r_snippets.md
@@ -1,0 +1,312 @@
+<!-- --8<-- [start:software] -->
+
+-   Unix terminal (e.g., `bash`)
+
+-   [Python environment with taxpasta](/#install)
+
+-   [R](https://rstudio-education.github.io/hopr/starting.html#how-to-download-and-install-r)
+
+-   Package dependencies
+
+    ```r
+    install.packages(c("readr", "dplyr"), dependencies = TRUE)
+    ```
+
+<!-- --8<-- [end:software] -->
+<!-- --8<-- [start:raw-motus] -->
+
+We can try loading a mOTUs profile into R using the common table reading
+function `read_tsv()` from the `readr` package with default arguments.
+
+```r
+requireNamespace("readr")
+```
+
+    Loading required namespace: readr
+
+```r
+profile_motus <- readr::read_tsv("2612_pe-ERR5766176-db_mOTU.out")
+```
+
+    Warning: One or more parsing issues, call `problems()` on your data frame for details,
+    e.g.:
+      dat <- vroom(...)
+      problems(dat)
+
+    Rows: 33573 Columns: 1
+
+    ── Column specification ────────────────────────────────────────────────────────
+    Delimiter: "\t"
+    chr (1): # git tag version 3.0.3 |  motus version 3.0.3 | map_tax 3.0.3 | ge...
+
+    ℹ Use `spec()` to retrieve the full column specification for this data.
+    ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
+
+You can see we immediately hit an error, as there is a ‘comment’ line at
+the top of the mOTUs profile with information on how the profile was
+generated.
+
+While such a comment is very useful for reproducibility, to load this we
+have to instead add extra options to the function, which makes loading
+the table less than smooth for downstream analyses.
+
+```r
+profile_motus <- readr::read_tsv("2612_pe-ERR5766176-db_mOTU.out", comment = "#")
+```
+
+    Warning: One or more parsing issues, call `problems()` on your data frame for details,
+    e.g.:
+      dat <- vroom(...)
+      problems(dat)
+
+    Rows: 33570 Columns: 3
+    ── Column specification ────────────────────────────────────────────────────────
+    Delimiter: "\t"
+    chr (1): Leptospira alexanderi [ref_mOTU_v3_00001]
+    dbl (2): 100053, 0
+
+    ℹ Use `spec()` to retrieve the full column specification for this data.
+    ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
+
+However, once again we hit another error: the column headers are _also_
+specified as a comment line… Instead we can try to skip the first two
+lines entirely.
+
+```r
+profile_motus <- readr::read_tsv("2612_pe-ERR5766176-db_mOTU.out", skip = 2)
+```
+
+    Rows: 33571 Columns: 3
+    ── Column specification ────────────────────────────────────────────────────────
+    Delimiter: "\t"
+    chr (1): #consensus_taxonomy
+    dbl (2): NCBI_tax_id, 2612_pe-ERR5766176-db_mOTU
+
+    ℹ Use `spec()` to retrieve the full column specification for this data.
+    ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
+
+```r
+profile_motus
+```
+
+    # A tibble: 33,571 × 3
+       `#consensus_taxonomy`                                      NCBI_tax…¹ 2612_…²
+       <chr>                                                           <dbl>   <dbl>
+     1 Leptospira alexanderi [ref_mOTU_v3_00001]                      100053       0
+     2 Leptospira weilii [ref_mOTU_v3_00002]                           28184       0
+     3 Chryseobacterium sp. [ref_mOTU_v3_00004]                           NA       0
[…]
