+++ .github/workflows/main.yml
@@ -30,19 +30,10 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Install dependencies
+      - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip setuptools wheel
           python -m pip install tox tox-gh-actions
-
-      - name: Install quarto & Test docs
-        if: (matrix.os == 'ubuntu-latest') && (matrix.python-version == '3.11')
-        run: |
-          curl -LO https://quarto.org/download/latest/quarto-linux-amd64.deb
-          sudo dpkg -i quarto-linux-amd64.deb
-          quarto check
-          Rscript -e "install.packages(c('rmarkdown', 'dplyr', 'readr'), dependencies = TRUE, repos = 'https://cran.rstudio.com')"
-          tox -e docs
 
       - name: Test with tox
         run: tox run-parallel --parallel auto --parallel-no-spinner -- --cov-report=xml
