+++ tox.ini
@@ -6,7 +6,7 @@ python =
     3.8: install, safety, py38
     3.9: install, safety, py39
     3.10: install, safety, py310
-    3.11: bandit, flake8, install, safety, mpypy, py311
+    3.11: bandit, flake8, safety, mypy, docs, install, py311
 
 [testenv]
 # Define base for all test environments.
@@ -54,11 +54,8 @@ commands=
 
 [testenv:docs]
 extras = dev
-allowlist_externals =
-    make
 commands=
     python -m bash_kernel.install
-    make docs
     mkdocs build --strict
 
 [testenv:install]
