### 2023-03-13 c69d0ad
chore: manually generate `make docs`
