+++ docs/tutorials/tutorial_bash_snippets.md
@@ -1,0 +1,179 @@
+<!-- --8<-- [start:data-setup] -->
+
+```bash
+mkdir taxpasta-tutorial
+cd taxpasta-tutorial
+```
+
+<!-- --8<-- [end:data-setup] -->
+<!-- --8<-- [start:data-curl] -->
+
+```bash
+## mOTUs
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
+
+## Kraken2
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
+```
+
+<!-- --8<-- [end:data-curl] -->
+<!-- --8<-- [start:data-wget] -->
+
+```bash
+## mOTUs
+wget https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
+wget https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
+
+## Kraken2
+wget https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
+```
+
+<!-- --8<-- [end:data-wget] -->
+<!-- --8<-- [start:data-ls] -->
+
+```bash
+ls
+```
+
+    2612_pe-ERR5766176-db1.kraken2.report.txt  2612_se-ERR5766180-db_mOTU.out
+
+    2612_pe-ERR5766176-db_mOTU.out
+
+<!-- --8<-- [end:data-ls] -->
+<!-- --8<-- [start:motus-head] -->
+
+```bash
+head 2612_pe-ERR5766176-db_mOTU.out
+```
+
+    # git tag version 3.0.3 |  motus version 3.0.3 | map_tax 3.0.3 | gene database: nr3.0.3 | calc_mgc 3.0.3 -y insert.scaled_counts -l 75 | calc_motu 3.0.3 -k mOTU -C no_CAMI -g 3 -c -p | taxonomy: ref_mOTU_3.0.3 meta_mOTU_3.0.3
+
+    # call: python /usr/local/bin/../share/motus-3.0.1//motus profile -p -c -f ERX5474932_ERR5766176_1.fastq.gz -r ERX5474932_ERR5766176_2.fastq.gz -db db_mOTU -t 2 -n 2612_pe-ERR5766176-db_mOTU -o 2612_pe-ERR5766176-db_mOTU.out
+
+    #consensus_taxonomy NCBI_tax_id 2612_pe-ERR5766176-db_mOTU
+
+    Leptospira alexanderi [ref_mOTU_v3_00001]   100053  0
+
+    Leptospira weilii [ref_mOTU_v3_00002]   28184   0
+
+    Chryseobacterium sp. [ref_mOTU_v3_00004]    NA  0
+
+    Chryseobacterium gallinarum [ref_mOTU_v3_00005] 1324352 0
+
+    Chryseobacterium indologenes [ref_mOTU_v3_00006]    253 0
+
+    Chryseobacterium artocarpi/ureilyticum [ref_mOTU_v3_00007]  NA  0
+
+    Chryseobacterium jejuense [ref_mOTU_v3_00008]   445960  0
+
[…]

[…]
