### 2023-03-13 0e7026e
chore: manual quarto, auto make docs
