### 2023-03-13 2e5a4f9
chore: manual quarto, auto make docs
