+++ tox.ini
@@ -54,8 +54,11 @@ commands=
 
 [testenv:docs]
 extras = dev
+allowlist_externals =
+    make
 commands=
     python -m bash_kernel.install
+    make docs
     mkdocs build --strict
 
 [testenv:install]
