### 2023-03-13 71b84a5
refactor: use bash snippets for how-tos
