### 2023-03-13 15b051d
refactor: use bash snippets for how-tos
