### 2023-03-13 baa02d7
chore: remove snippet generation since we check in results
