### 2023-03-13 fdb8089
chore: remove snippet generation since we check in results
