+++ .readthedocs.yaml
@@ -6,10 +6,6 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.10"
-  jobs:
-    pre_build:
-      - python -m bash_kernel.install
-      - make docs
 
 mkdocs:
   configuration: mkdocs.yml
