### 2023-03-13 7d9695a
refactor: remove the consensus command until implementation
