+++ src/taxpasta/infrastructure/cli/consensus.py
@@ -1,28 +1,0 @@
-# Copyright (c) 2022 Moritz E. Beber
-# Copyright (c) 2022 Maxime Borry
-# Copyright (c) 2022 James A. Fellows Yates
-# Copyright (c) 2022 Sofia Stamouli.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-"""Add the `consensus` command to the taxpasta CLI."""
-
-
-from .taxpasta import app
-
-
-@app.command()
-def consensus():
-    """Form a consensus for the same sample but from different taxonomic profiles."""
-    raise NotImplementedError("Coming soon™")
