+++ docs/quick_reference/consensus.md
@@ -1,13 +1,0 @@
-# consensus
-
-!!! failure
-
-    Not yet implemented!
-
-```shell
-taxpasta consensus --help
-```
-
-```text
---8<-- "consensus_help.txt"
-```
