+++ docs/commands/index.md
@@ -6,4 +6,3 @@ The following pages document the 'what, how, and why' for each command.
 
 -   [`taxpasta standardise`](standardise.md)
 -   [`taxpasta merge`](merge.md)
--   [`taxpasta consensus`](consensus.md)
