+++ docs/quick_reference/index.md
@@ -4,4 +4,3 @@ The following pages provide the `help` messages of each command to act as a quic
 
 -   [`taxpasta standardise`](standardise.md)
 -   [`taxpasta merge`](merge.md)
--   [`taxpasta consensus`](consensus.md)
