+++ docs/commands/consensus.md
@@ -1,5 +1,0 @@
-# consensus
-
-!!! failure
-
-    Not yet implemented!
