### 2023-03-13 689bcef
refactor: remove the consensus command until implementation
