+++ mkdocs.yml
@@ -31,6 +31,7 @@ markdown_extensions:
 theme:
   features:
     - content.code.annotate
+    - content.code.copy
     - navigation.indexes
     - navigation.instant
     - navigation.tracking
