### 2023-03-13 0bcf4b2
docs: expand the usage section
