+++ README.md
@@ -25,6 +25,24 @@ for the same profiler into a single table. In future, we also intend to offer
 methods for forming a _consensus_ for the same sample analyzed by different
 profilers.
 
+### Supported Taxonomic Profilers
+
+Taxpasta currently supports standardisation and generation of comparable
+taxonomic tables for:
+
+-   [Bracken](https://ccb.jhu.edu/software/bracken/)
+-   [Centrifuge](https://ccb.jhu.edu/software/centrifuge/)
+-   [DIAMOND](https://github.com/bbuchfink/diamond)
+-   [Kaiju](https://kaiju.binf.ku.dk/)
+-   [Kraken2](https://ccb.jhu.edu/software/kraken2/)
+-   [KrakenUniq](https://github.com/fbreitwieser/krakenuniq)
+-   [MEGAN6](http://www-ab.informatik.uni-tuebingen.de/software/megan6)/[MALT](https://uni-tuebingen.de/fakultaeten/mathematisch-naturwissenschaftliche-fakultaet/fachbereiche/informatik/lehrstuehle/algorithms-in-bioinformatics/software/malt/)
+-   [MetaPhlAn](https://segatalab.cibio.unitn.it/tools/metaphlan/index.html)
+-   [mOTUs](https://motu-tool.org/)
+
+See [supported profilers](https://taxpasta.readthedocs.io/en/latest/supported_profilers/)
+for more information.
+
 ## Install
 
 It's as simple as:
@@ -63,6 +81,63 @@ pip install 'taxpasta[rich,biom]'
 -   `xlsx` supports writing output tables in [Microsoft Excel](https://support.microsoft.com/en-us/office/file-formats-that-are-supported-in-excel-0943ff2c-6014-4e8d-aaea-b83d51d46247) format.
 -   `all` includes all of the above.
 -   `dev` provides all tools needed for contributing to taxpasta.
+
+## Usage
+
+The main entry point for taxpasta is its command-line interface (CLI). You can interactively
+explore the offered commands through the help system.
+
+```shell
+taxpasta -h
+```
+
+Taxpasta currently offers two commands corresponding to the main
+[use-cases](#about). You can find out more in the [commands'
+documentation](https://taxpasta.readthedocs.io/en/latest/commands/).
+
+### Standardise
+
+Since the [supported profilers](#supported-taxonomic-profilers) all produce
+their own flavour of tabular output, a quick way to normalize such files, is to
+standardise them with taxpasta. You need to let taxpasta know what tool the file
+was created by. As an example, let's standardise a MetaPhlAn profile. (You can
+find an example file in our [test
+data](https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/metaphlan/MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt).)
+
+```shell
+taxpasta standardise -p metaphlan -o standardised.tsv MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt
+```
+
+With these minimal arguments, taxpasta produces a two column output consisting of
+
+| taxonomy_id | count |
+| ----------- | ----- |
+|             |       |
+
+You can count on the second column being integers :wink: Having such a simple
+and tidy table should make your downstream analysis much smoother to start out
+with. Please, have a look at the full [getting
+started](https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started.md)
+tutorial for a more thorough introduction.
+
+### Merge
+
+Converting single tables is nice, but hopefully you have many shiny samples to
+analyze. The `taxpasta merge` command works similarly to `standardise` except
[…]
