### 2023-03-13 f086fa7
docs: expand the usage section
