### 2023-03-13 1bdac26
docs: turn note admonitions into info ones
