### 2023-03-13 cc8cd2a
docs: turn note admonitions into info ones
