+++ docs/supported_profilers/index.md
@@ -1,6 +1,15 @@
 # Supported Tools
 
-The following pages provide more information on the types of output files of each taxonomic profiler can be standardised by `taxpasta`.
+The following pages provide more information on the types of output files of
+each taxonomic profiler can be standardised by taxpasta.
+
+!!! warning
+
+    In particular, when comparing results between various tools, please be very
+    aware of their differences.  In the document on
+    [terminology](terminology.md), we describe how we think about the different
+    tools.  Spending some time on considering the consequences now, will save
+    you a lot of headache and prevent potentially wrong results later.
 
 -   [Bracken](bracken.md)
 -   [Centrifuge](centrifuge.md)
