### 2023-03-13 43b6041
docs: shorten command summary
