+++ docs/tutorials/tutorial_python_snippets.qmd
@@ -144,7 +144,7 @@ print(profile_kraken2.head().to_markdown())
 With pandas, we can perform such a merge either with the `join` or `merge`
 methods and an `"outer"` argument to perform an outer join that includes all
 rows both from the left and right table in the resulting table. The `join`
-method works best when joining tables on their indeces. In our case, the indeces
+method works best when joining tables on their indices. In our case, the indices
 are not meaningful for the join operation as they are simple integer ranges. We
 will use the `merge` method instead.
 
