### 2023-03-13 4ea443e
Apply suggestions from code review
Co-authored-by: James A. Fellows Yates <jfy133@gmail.com>
