### 2023-03-13 17d64ca
Apply suggestions from code review
Co-authored-by: James A. Fellows Yates <jfy133@gmail.com>
