+++ docs/supported_profilers/index.md
@@ -8,7 +8,7 @@ each taxonomic profiler can be standardised by taxpasta.
     In particular, when comparing results between various tools, please be very
     aware of their differences.  In the document on
     [terminology](terminology.md), we describe how we think about the different
-    tools.  Spending some time on considering the consequences now, will save
+    tools. Spending some time on considering the consequences now, will save
     you a lot of headache and prevent potentially wrong results later.
 
 -   [Bracken](bracken.md)
