### 2023-03-13 c55353f
docs: generate snippets without rich
