+++ docs/quick_reference/standardise_help.txt
@@ -1,78 +1,45 @@
-                                                                                              
- Usage: taxpasta standardise [OPTIONS] PROFILE                                                
-                                                                                              
- Standardise a taxonomic profile (alias: 'standardize').                                      
-                                                                                              
-╭─ Arguments ────────────────────────────────────────────────────────────────────────────────╮
-│ *    profile      PATH  A file containing a taxonomic profile. [required]                  │
-╰────────────────────────────────────────────────────────────────────────────────────────────╯
-╭─ Options ──────────────────────────────────────────────────────────────────────────────────╮
-│ *  --profiler                 -p      [bracken|centrifuge|diamo  The taxonomic profiler    │
-│                                       nd|kaiju|kraken2|krakenun  used.                     │
-│                                       iq|megan6|metaphlan|motus  [required]                │
-│                                       ]                                                    │
-│ *  --output                   -o      PATH                       The desired output file.  │
-│                                                                  By default, the file      │
-│                                                                  extension will be used to │
-│                                                                  determine the output      │
-│                                                                  format.                   │
-│                                                                  [required]                │
-│    --output-format                    [TSV|CSV|ODS|XLSX|arrow|p  The desired output        │
-│                                       arquet]                    format. Depending on the  │
-│                                                                  choice, additional        │
-│                                                                  package dependencies may  │
-│                                                                  apply. Will be parsed     │
-│                                                                  from the output file name │
-│                                                                  but can be set            │
-│                                                                  explicitly.               │
-│                                                                  [default: None]           │
-│    --summarise-at,--summari…          TEXT                       Summarise abundance       │
-│                                                                  profiles at higher        │
-│                                                                  taxonomic rank. The       │
-│                                                                  provided option must      │
-│                                                                  match a rank in the       │
-│                                                                  taxonomy exactly. This is │
-│                                                                  akin to the clade         │
-│                                                                  assigned reads provided   │
-│                                                                  by, for example, kraken2, │
-│                                                                  where the abundances of a │
-│                                                                  whole taxonomic branch    │
-│                                                                  are assigned to a taxon   │
-│                                                                  at the desired rank.      │
-│                                                                  Please note that          │
-│                                                                  abundances above the      │
-│                                                                  selected rank are simply  │
-│                                                                  ignored. No attempt is    │
-│                                                                  made to redistribute      │
-│                                                                  those down to the desired │
-│                                                                  rank. Some tools, like    │
-│                                                                  Bracken, were designed    │
-│                                                                  for this purpose but it   │
-│                                                                  doesn't seem like a       │
-│                                                                  problem we can generally  │
-│                                                                  solve here.               │
-│                                                                  [default: None]           │
-│    --taxonomy                         PATH                       The path to a directory   │
-│                                                                  containing taxdump files. │
-│                                                                  At least nodes.dmp and    │
-│                                                                  names.dmp are required. A │
-│                                                                  merged.dmp file is        │
-│                                                                  optional.                 │
-│                                                                  [default: None]           │
-│    --add-name                                                    Add the taxon name to the │
-│                                                                  output.                   │
-│    --add-rank                                                    Add the taxon rank to the │
-│                                                                  output.                   │
-│    --add-lineage                                                 Add the taxon's entire    │
-│                                                                  lineage to the output.    │
[…]
