### 2023-03-13 7725171
docs: generate snippets without rich
