### 2023-03-13 #84 @Midnighter
Dear @NickleDave,

We think that we've implemented everything that you've noticed. Please take another look and let us know your thoughts.

* There's now a detailed [tutorial](https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started/)
* We've shortened the command summaries so hopefully they won't be truncated (except on extremely small windows).
* There's a short usage section [right in the readme](https://github.com/taxprofiler/taxpasta#usage) and on the docs' [main page](https://taxpasta.readthedocs.io/en/latest/#usage).
* [Extras](https://github.com/taxprofiler/taxpasta#optional-dependencies) are documented.
* A number of other improvements to the documentation.
