### 2023-03-23 #84 @ctb
> I guess we'll have to add support for sourmash then 😆 Good to hear and I look forward to your review @ctb. Thank you for your time.

welcome! ironically we have invested time in generating the same report formats as you, for other consumers of taxonomy, so I think we are well positioned to engage productively ;).
