### 2023-03-31 #84 @Midnighter
I have a question about the version submitted. We've already made new releases since then and I'm planning another round of smaller fixes. It'd make sense to me that the latest version will be reviewed rather than what we started with. Should I just edit the original post or is that against your procedure?
