### 2023-04-01 82dc0e8
refactor: generally allow 1% tolerance of compositionality
* Allow 2% for Bracken.
