### 2023-04-01 fbcfc9d
refactor: generally allow 1% tolerance of compositionality
* Allow 2% for Bracken.
