+++ src/taxpasta/infrastructure/application/centrifuge/centrifuge_profile.py
@@ -39,9 +39,7 @@ class CentrifugeProfile(pa.SchemaModel):
     @classmethod
     def check_compositionality(cls, percent: Series[float]) -> bool:
         """Check that the percent of 'unclassified' and 'root' add up to a hundred."""
-        return len(percent) == 0 or bool(
-            np.isclose(percent[:2].sum(), 100.0, atol=0.01)
-        )
+        return percent.empty or bool(np.isclose(percent[:2].sum(), 100.0, atol=1.0))
 
     class Config:
         """Configure the schema model."""
