### 2023-04-01 691de1a
chore: increase minimum pandera version
