### 2023-04-01 a2495bb
refactor: use `DataFrameModel` instead of `SchemaModel`
