### 2023-04-01 b6da975
refactor: use `DataFrameModel` instead of `SchemaModel`
