### 2023-04-01 3034d8d
fix: avoid int32 type failure on Windows
