### 2023-04-01 553f926
fix: avoid int32 type failure on Windows
