### 2023-04-01 46bf412
fix: cast types to numpy.int64 explicitly
