### 2023-04-01 cd20f64
fix: handle tax IDs being -1 and combine unclassified
