### 2023-04-01 0606f04
refactor: use integer with missing values
