+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/arrow_wide_observation_table_writer.py
@@ -36,7 +36,7 @@ class ArrowWideObservationTableWriter(WideObservationTableWriter):
         cls,
         matrix: DataFrame[WideObservationTable],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given table to the given buffer or file."""
         matrix.to_feather(target, **kwargs)
