+++ src/taxpasta/infrastructure/application/standard_profile_writer/xlsx_standard_profile_writer.py
@@ -33,7 +33,7 @@ class XLSXStandardProfileWriter(StandardProfileWriter):
         cls,
         profile: DataFrame[StandardProfile],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given standardized profile to the given buffer or file."""
         profile.to_excel(target, index=False, engine="openpyxl", **kwargs)
