+++ src/taxpasta/infrastructure/application/standard_profile_writer/ods_standard_profile_writer.py
@@ -33,7 +33,7 @@ class ODSStandardProfileWriter(StandardProfileWriter):
         cls,
         profile: DataFrame[StandardProfile],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given standardized profile to the given buffer or file."""
         profile.to_excel(target, index=False, engine="odf", **kwargs)
