+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/ods_wide_observation_table_writer.py
@@ -36,7 +36,7 @@ class ODSWideObservationTableWriter(WideObservationTableWriter):
         cls,
         matrix: DataFrame[WideObservationTable],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given table to the given buffer or file."""
         matrix.to_excel(target, index=False, engine="odf", **kwargs)
