+++ src/taxpasta/infrastructure/application/tidy_observation_table_writer/xlsx_table_writer.py
@@ -36,7 +36,7 @@ class XLSXTidyObservationTableWriter(TidyObservationTableWriter):
         cls,
         table: DataFrame[TidyObservationTable],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given table to the given buffer or file."""
         table.to_excel(target, index=False, engine="openpyxl", **kwargs)
