+++ src/taxpasta/infrastructure/application/tidy_observation_table_writer/arrow_table_writer.py
@@ -36,7 +36,7 @@ class ArrowTidyObservationTableWriter(TidyObservationTableWriter):
         cls,
         table: DataFrame[TidyObservationTable],
         target: BinaryBufferOrFilepath,
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given table to the given buffer or file."""
         table.to_feather(target, **kwargs)
