+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/biom_wide_observation_table_writer.py
@@ -41,7 +41,7 @@ class BIOMWideObservationTableWriter(WideObservationTableWriter):
         target: Filepath,
         taxonomy: Optional[TaxonomyService] = None,
         generated_by: str = "taxpasta",
-        **kwargs
+        **kwargs,
     ) -> None:
         """Write the given data to the given buffer or file."""
         # Drop unclassified reads.
