### 2023-04-01 321c178
fix: correctly address column
Cannot use generic type to get attribute.
