### 2023-04-20 #84 @snacktavish
OK - I am revealing my slow start, but as far as I can tell, the tutorial is not rendering correctly https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started.md, and it's not really follow-able at https://github.com/taxprofiler/taxpasta/blob/dev/docs/tutorials/getting-started.md
Is there a better link I should be using?

Thanks,
Emily Jane
