### 2023-04-21 #84 @jfy133
Hi @snacktavish !

No worries :).

It think we've made a link in the website with the wrong extension:

https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started/

Should work. I'll try find all instances of that mistake and fix them.

However I'mn not sure why it doens't render at all on the dev branch, I'll look into that now. But I don't think there is much difference in that file with the master branch.

Cheers,
