+++ docs/supported_profilers/index.md
@@ -7,16 +7,16 @@ each taxonomic profiler can be standardised by taxpasta.
 
     In particular, when comparing results between various tools, please be very
     aware of their differences.  In the document on
-    [terminology](terminology.md), we describe how we think about the different
+    [terminology](terminology), we describe how we think about the different
     tools. Spending some time on considering the consequences now, will save
     you a lot of headache and prevent potentially wrong results later.
 
--   [Bracken](bracken.md)
--   [Centrifuge](centrifuge.md)
--   [DIAMOND](diamond.md)
--   [Kaiju](kaiju.md)
--   [Kraken2](kraken2.md)
--   [KrakenUniq](krakenuniq.md)
--   [MEGAN6/MALT](megan6.md)
--   [MetaPhlAn](metaphlan.md)
--   [mOTUs](motus.md)
+-   [Bracken](bracken)
+-   [Centrifuge](centrifuge)
+-   [DIAMOND](diamond)
+-   [Kaiju](kaiju)
+-   [Kraken2](kraken2)
+-   [KrakenUniq](krakenuniq)
+-   [MEGAN6/MALT](megan6)
+-   [MetaPhlAn](metaphlan)
+-   [mOTUs](motus)
