### 2023-04-21 taxprofiler/taxpasta#88 @jfy133
[BUG] Some internal links on readthedeocs broken
### Is there an existing issue for this?

- [X] I have searched the existing issues

### Problem description

https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1517061407

### Code sample

Code run:

Traceback:

### Anything else?

_No response_
