+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed a few broken internal links (#89)
+
 ## [0.3.0] - (2023-04-01)
 
 ### Added
