### 2023-04-22 #84 @bluegenes
Hi folks, here's my review.

Great tool - having tried to provide some standardized formats for our profiler, I definitely appreciate the utility provided here!

    - _I see Contributing info in the documentation, but it would be good to link to this page from the README_

        - _I see a pypi version badge, but not conda. table displays badges nicely._

        - _yes, command usage examples exist for both `merge` and `standardise`, but fully functional examples with file download links would be much better_
        -  in the `Usage` section, `use-cases` just links to the `about` section higher in the README.
        - in the `Merge` section, test-data` links to a single file example. Two files are needed for the `merge` command.
    - I don't see any citation info

Estimated hours spent reviewing: 4-5

---

The `taxpasta` software tackles an important issue in metagenomic taxonomic profiling by ingesting the output from a number of popular profilers and producing standardized outputs that can be used in downstream workflows.

## Tutorial comments

I find the back-and-forth between CLI and R/python distracting in the tutorial -- It would be clearer to see the entire command line + execution for `taxpasta`, with brief explanations around each command together, since the tool is  CLI-based and the other sections are just demonstrating the justification for the tool.

In a "Getting Started" tutorial, including code for the failures (including partial solutions) of standard reading methods might confuse folks who are newer to coding/bioinfo. People may end up confusing which lines they should run on their own data and which they shouldn't.

I think the R/python parts are more appropriate for a follow-up section or tutorial, entitled "Why taxpasta" or similar to indicate that you're digging into the rationale and issues that prompted building taxpasta. That way, beginners who just want to run the cli are comfortable, and intermediate/etc can dig in afterwards. Note, I only tested the python code parts (not R), but those worked great.

If kept as is, tutorial at least needs a bolded line calling attention to the switch between command line and R/Python sections, -- e.g.
> **The following section uses R/Python.**

and similar when you switch back.

I think you're already be aware, but several links are broken from the `Getting Started` tutorial:

- https://taxpasta.readthedocs.io/how-tos/how-to-add-names/
    - should be: https://taxpasta.readthedocs.io/en/latest/how-tos/how-to-add-names/
- https://taxpasta.readthedocs.io/how-tos/how-to-customise-sample-names/
- https://taxpasta.readthedocs.io/supported_profilers/terminology

The `adding taxa names` how-to tutorial could usefully be directly added to the main `Getting Started` cli tutorial if you switch up the bash/python structure (I do see it linked at the bottom, though!).

## Commands and Documentation

- In the `README`, you include a full command for `standardise` and state that an example file can be found in test data. Can you instead include small test files with the package or add a `curl` command to download the exact file, so that command is actually executable? Same for the next command in the README (add `curl` to dl a second file to allow `merge`). Same comment for documentation pages for these.
- There were several screens of terminal output upon command failure (specifically, running the `standardise` README command without downloading the file). Might be confusing for new folks (even though it's very pretty! :) and the only line that matters is the very last one,
`FileNotFoundError: [Errno 2] No such file or directory: 'MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt` Might be out of scope, just wanted to mention.
- It would be helpful to add a general note in the README that users can see additional tool help via `taxpasta <command> -h`, e.g. `taxpasta merge -h`.
- It would be nice to note somewhere that setting output format explicitly overrides use of the file extension for determining output format (I noticed I could write a `arrow` format with `.tsv` extension)
- Definitely out of scope, but wanted to suggest an extension: have you considered adding a `filter` functionality, e.g. by % of total dataset reported in the file? There's some evidence that filtering e.g. Kraken2 output might help reduce false positives (e.g. [here](https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-022-05103-0) )
