### 2023-04-22 taxprofiler/taxpasta#90 @Midnighter
Add repostatus.org badge
Based on review at https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1518457423
