### 2023-04-22 taxprofiler/taxpasta#92 @Midnighter
Document a curl command to download sample files in the readme
Based on https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1518457423
