### 2023-04-22 taxprofiler/taxpasta#93 @Midnighter
Document that setting the output format overrules any given file extension
Based on https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1518457423
