### 2023-04-22 #84 @Midnighter
Dear @bluegenes,

Thank you very much for your kind review. I'll try to respond to all your points now and note where we will take action.

> A [repostatus.org](https://www.repostatus.org/) badge,

We're [adding a repo status badge](https://github.com/taxprofiler/taxpasta/issues/90)

> If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.

Agreed that this would be a good addition. I think we will write a JOSS paper after seeing the second review and copy this kind of information from there into the documentation.

> I don't see any citation info

Technically, the Zenodo DOI already makes this citeable but we'll add more info if taxpasta gets accepted here and at JOSS.

> JOSS's requirements

As mentioned above, we plan to write the paper after the initial review.

> Tutorial comments

Thank you for the detailed observations on the tutorial. We agree with your points and @jfy133 will work on splitting this up and improving the getting started tutorial. He [already fixed the URLs](https://github.com/taxprofiler/taxpasta/pull/89) that were wrong.

> In the README, you include a full command for standardise and state that an example file can be found in test data. Can you instead include small test files with the package or add a curl command to download the exact file, so that command is actually executable? Same for the next command in the README (add curl to dl a second file to allow merge). Same comment for documentation pages for these.

Good point, [we will do so](https://github.com/taxprofiler/taxpasta/issues/92).

> There were several screens of terminal output upon command failure (specifically, running the standardise README command without downloading the file). Might be confusing for new folks (even though it's very pretty! :) and the only line that matters is the very last one,
FileNotFoundError: [Errno 2] No such file or directory: 'MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt Might be out of scope, just wanted to mention.

We will consider this when looking at the tutorial and the readme for your other points above.

> It would be helpful to add a general note in the README that users can see additional tool help via taxpasta <command> -h, e.g. taxpasta merge -h.

We thought that we had covered this in the [Usage section](https://github.com/taxprofiler/taxpasta#usage) of the readme. Do you think that it needs to be stated there that this also works for sub commands?

> It would be nice to note somewhere that setting output format explicitly overrides use of the file extension for determining output format (I noticed I could write a arrow format with .tsv extension)

[Good call](https://github.com/taxprofiler/taxpasta/issues/93).

> Definitely out of scope, but wanted to suggest an extension: have you considered adding a filter functionality, e.g. by % of total dataset reported in the file? T[here](https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-022-05103-0)'s some evidence that filtering e.g. Kraken2 output might help reduce false positives (e.g. here )

We agree that, although this is useful in general, it is out of scope for taxpasta and should happen downstream. We rather follow the [UNIX philosophy](https://en.wikipedia.org/wiki/Unix_philosophy#Origin) here.

> Make each program do one thing well. To do a new job, build afresh rather than complicate old programs by adding new "features".

Thus, we can see a case for two new packages taxpasta-stats and taxpasta-viz in the future that take advantage of these standardised tables but haven't done any scoping work or thought about all the desired capabilities yet.
