+++ README.md
@@ -40,7 +40,7 @@ taxonomic tables for:
 -   [MetaPhlAn](https://segatalab.cibio.unitn.it/tools/metaphlan/index.html)
 -   [mOTUs](https://motu-tool.org/)
 
-See [supported profilers](/supported_profilers/)
+See [supported profilers](https://taxpasta.readthedocs.io/en/latest/supported_profilers/)
 for more information.
 
 ## Install
@@ -93,7 +93,7 @@ taxpasta -h
 
 Taxpasta currently offers two commands corresponding to the main
 [use-cases](#about). You can find out more in the [commands'
-documentation](/commands).
+documentation](https://taxpasta.readthedocs.io/en/latest/commands).
 
 ### Standardise
 
@@ -117,7 +117,7 @@ With these minimal arguments, taxpasta produces a two column output consisting o
 You can count on the second column being integers :wink:. Having such a simple
 and tidy table should make your downstream analysis much smoother to start out
 with. Please have a look at the full [getting
-started](/tutorials/getting-started)
+started](https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started)
 tutorial for a more thorough introduction.
 
 ### Merge
@@ -136,7 +136,7 @@ taxpasta merge -p metaphlan -o merged.tsv MOCK_*.metaphlan3_profile.txt
 The output of the `merge` command has one column for the taxonomic identifier and
 one more column for each input profile. Again, have a look at the full
 [getting
-started](/tutorials/getting-started)
+started](https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started)
 tutorial for a more thorough introduction.
 
 ## Copyright
