### 2023-04-25 taxprofiler/taxpasta#94 @jfy133
[BUG] Getting started tutorial is too complex
### Is there an existing issue for this?

- [X] I have searched the existing issues

### Problem description

x-ref https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1518457423

Should simplify to just main workflow

Current getting started should move to a 'deep dive' type tutorial

### Code sample

_No response_

### Environment

NA

### Anything else?

_No response_
