### 2023-04-25 36f5fdc
Split up getting started to simplified version and deep dive
