+++ docs/tutorials/getting-started.md
@@ -1,42 +1,20 @@
+
 # Getting Started
 
-## Introduction
+In this getting started tutorial we will show you how to generate
+standardised taxonomic profiles from thediverse outputs of two
+popular taxonomic profilers: [Kraken2](https://ccb.jhu.edu/software/kraken2/)
+and [mOTUs](https://motu-tool.org/) using `taxpasta`.
 
-We will show you how to generate standardised taxonomic profiles from the
-diverse outputs of two popular taxonomic profilers:
-[Kraken2](https://ccb.jhu.edu/software/kraken2/) and
-[mOTUs](https://motu-tool.org/). We demonstrate some benefits of this
-standardisation when running downstream analyses on such tables in either the
-popular statistical programming language [R](https://www.r-project.org/) or
-[Python](https://www.python.org/).
+If you want a more detailed walkthrough of _why_ standardising the profiles
+are useful, please see the [Deep Dive](/tutorials/deepdive) tutorial.
 
 ## Preparation
 
 ### Software
 
 For this tutorial you will need an internet connection, an [installation of
-taxpasta](/#install), and, if you want to follow the R parts, an [installation
-of
-R](https://rstudio-education.github.io/hopr/starting.html#how-to-download-and-install-r)
-itself with the [readr](https://readr.tidyverse.org/) and
-[dplyr](https://dplyr.tidyverse.org/) packages from the
-[Tidyverse](https://tidyverse.org). Furthermore, we assume a UNIX based
-operating system, such as Linux, macOS, or Windows Subsystem for Linux.
-
-To summarise, you will need:
-
-
-=== "R"
-
---8<--
-    tutorial_r_snippets.md:software
---8<--
-
-=== "Python"
-
---8<--
-    tutorial_python_snippets.md:software
---8<--
+taxpasta](/#install).
 
 ### Data
 
@@ -49,7 +27,7 @@ tutorial_bash_snippets.md:data-setup
 
 We will also need to download some example taxonomic profiles from Kraken2 and
 mOTUs. We can download test data from the taxpasta repository using, for
-example, `curl` or `wget`.
+example, `curl` (OSX, Linux) or `wget` (generally Linux Only).
 
 === "curl"
 
@@ -87,76 +65,11 @@ To begin, let’s look at the contents of the output from each profiler.
     tutorial_bash_snippets.md:kraken2-head
 --8<--
 
-These look quite different and neither of them is in a nice "pure" tabular
-format that is convenient for analysis software to load. They also have
-different types columns and, in the case of Kraken2, it has an interesting
+These look quite different. Neither of them is in a nice "pure" tabular
+format that is convenient for analysis software or spreadsheet tools such
+as Microsoft Excel or LibreOffice Calc to load. They also have different
+types columns and, in the case of Kraken2, it has an interesting
 "indentation" way of showing the taxonomic rank of each taxon.
-
-#### mOTUs
-
-=== "R"
-
---8<--
-    tutorial_r_snippets.md:raw-motus
---8<--
-
-=== "Python"
-
---8<--
-    tutorial_python_snippets.md:raw-motus
---8<--
-
-That works! At least in terms of reasonable table headers. However, we can see
-that there are missing NCBI taxonomy identifiers so there is probably more data
-cleaning ahead of us.  Getting to this point took too much effort already to
-load what is essentially a simple table. Furthermore, if we are interested in
-loading all mOTUs profiles, we would have to load each profile one by one for
-each sample, requiring more complicated loops and table join code.
-
-#### Kraken2
-
-And what about the Kraken2 output?
-
-=== "R"
-
---8<--
-    tutorial_r_snippets.md:raw-kraken2
---8<--
-
[…]
