### 2023-04-25 #84 @jfy133
Hi @bluegenes, your comments on the Tutorial have (hopeully) been addressed here: https://github.com/taxprofiler/taxpasta/pull/95 once merged!
