### 2023-04-27 b3388fb
Apply suggestions from code review
Co-authored-by: Moritz E. Beber <midnighter@posteo.net>
