### 2023-04-29 048c892
docs: improve description of options behavior
fix #93
