### 2023-04-29 4915083
docs: improve description of options behavior
fix #93
