### 2023-04-29 0bdf22e
docs: drop mention of consensus command
