### 2023-04-29 308ba7f
docs: drop mention of consensus command
