+++ docs/index.md
@@ -12,12 +12,8 @@ as well as any number of additional columns with extra information. We therefore
 decided to take [the lessons learnt](https://xkcd.com/927/) to heart and provide
 our own solution to deal with this pasticcio. With taxpasta you can ingest all
 of those formats and, at a minimum, output taxonomy identifiers and their
-integer counts.
-
-Taxpasta can not only standardise profiles but also _merge_ them across samples
-for the same profiler into a single table. In future, we also intend to offer
-methods for forming a _consensus_ for the same sample analyzed by different
-profilers.
+integer counts. Taxpasta can not only standardise profiles but also _merge_ them
+across samples for the same profiler into a single table.
 
 ### Supported Taxonomic Profilers
 
