### 2023-04-29 01f8eac
docs: add commands for downloading sample files
Use the more stable `main` branch to fetch example files.
