+++ docs/index.md
@@ -92,9 +92,11 @@ their own flavour of tabular output, a quick way to normalize such files, is to
 standardise them with taxpasta. You need to let taxpasta know what tool the file
 was created by. As an example, let's standardise a MetaPhlAn profile. (You can
 find an example file in our [test
-data](https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/metaphlan/MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt).)
+data](https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/metaphlan/MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt).)
 
 ```shell
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/metaphlan/MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt
+
 taxpasta standardise -p metaphlan -o standardised.tsv MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt
 ```
 
@@ -116,10 +118,15 @@ Converting single tables is nice, but hopefully you have many shiny samples to
 analyze. The `taxpasta merge` command works similarly to `standardise` except
 that you provide multiple profiles as input. Grab a few more MOCK examples from
 our [test
-data](https://github.com/taxprofiler/taxpasta/tree/dev/tests/data/metaphlan) and
+data](https://github.com/taxprofiler/taxpasta/tree/main/tests/data/metaphlan) and
 try it out.
 
 ```shell
+LOCATION=https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/metaphlan
+curl -O "${LOCATION}/MOCK_001_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt"
+curl -O "${LOCATION}/MOCK_002_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt"
+curl -O "${LOCATION}/MOCK_003_Illumina_Hiseq_3000_se_metaphlan3-db.metaphlan3_profile.txt"
+
 taxpasta merge -p metaphlan -o merged.tsv MOCK_*.metaphlan3_profile.txt
 ```
 
