### 2023-04-29 cd15aa4
docs: add commands for downloading sample files
Use the more stable `main` branch to fetch example files.
