### 2023-04-29 f6e500f
fix: remove non-existant setup.py reference
