### 2023-04-29 6c2f8b0
docs: document quarto and bash-kernel installation
