+++ docs/how-tos/bash_snippets.md
@@ -5,11 +5,11 @@ cd taxpasta-tutorial
 
 ```bash
 ## mOTUs
-curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
-curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/motus/2612_pe-ERR5766176-db_mOTU.out
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/motus/2612_se-ERR5766180-db_mOTU.out
 
 ## Kraken2
-curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/dev/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
+curl -O https://raw.githubusercontent.com/taxprofiler/taxpasta/main/tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
 ```
 
 <!-- --8<-- [start:taxdump] -->
@@ -23,37 +23,11 @@ tar -C taxdump -xzf taxdump.tar.gz
 ```
 
       % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-
                                      Dload  Upload   Total   Spent    Left  Speed
-
-      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-
-      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-
-    100    49  100    49    0     0     38      0  0:00:01  0:00:01 --:--:--    38
-
+      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0100    49  100    49    0     0     33      0  0:00:01  0:00:01 --:--:--    33
       % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-
                                      Dload  Upload   Total   Spent    Left  Speed
-
-      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-
-      0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
-
-      0 57.8M    0 17376    0     0  13692      0  1:13:49  0:00:01  1:13:48 13681
-
-      8 57.8M    8 4812k    0     0  2169k      0  0:00:27  0:00:02  0:00:25 2168k
-
-     25 57.8M   25 14.8M    0     0  4718k      0  0:00:12  0:00:03  0:00:09 4718k
-
-     42 57.8M   42 24.8M    0     0  6024k      0  0:00:09  0:00:04  0:00:05 6022k
-
-     62 57.8M   62 36.3M    0     0  7159k      0  0:00:08  0:00:05  0:00:03 7447k
-
-     85 57.8M   85 49.2M    0     0  8093k      0  0:00:07  0:00:06  0:00:01  9.9M
-
-    100 57.8M  100 57.8M    0     0  8566k      0  0:00:06  0:00:06 --:--:-- 11.3M
-
[…]
