### 2023-04-29 2d16176
chore: generate command help output
