+++ docs/quick_reference/standardise_help.txt
@@ -10,12 +10,16 @@ Options:
                                   The taxonomic profiler used.  [required]
   -o, --output PATH               The desired output file. By default, the
                                   file extension will be used to determine the
-                                  output format.  [required]
+                                  output format, but when setting the format
+                                  explicitly using the --output-format option,
+                                  automatic detection is disabled.  [required]
   --output-format [TSV|CSV|ODS|XLSX|arrow|parquet]
                                   The desired output format. Depending on the
                                   choice, additional package dependencies may
-                                  apply. Will be parsed from the output file
-                                  name but can be set explicitly.
+                                  apply. By default it will be parsed from the
+                                  output file name but it can be set
+                                  explicitly and will then disable the
+                                  automatic detection.
   --summarise-at, --summarize-at TEXT
                                   Summarise abundance profiles at higher
                                   taxonomic rank. The provided option must
