### 2023-05-02 taxprofiler/taxpasta#98 @jfy133
[BUG] Improve docs on tutorial data & warning against cross-profile merging
### Is there an existing issue for this?

- [X] I have searched the existing issues

### Problem description

Based on reviewer here https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1530191450

To do:

- [x] Provide more biological context on example data
- [x] Add more docs woranings about merging across profilers
- [x] Give examples of how to merge across profilers if a user really wants
