### 2023-05-02 #84 @Midnighter
Dear @snacktavish,

Thank you for your review. As a quick comment before answering in more detail, where did you find the URL to the how-to that is not found? The proper link is https://taxpasta.readthedocs.io/en/latest/how-tos/how-to-add-names/
