### 2023-05-02 24a5e50
Apply suggestions from code review
Co-authored-by: Sofia Stamouli <91951607+sofstam@users.noreply.github.com>
