### 2023-05-02 d193eff
Apply suggestions from code review
Co-authored-by: Sofia Stamouli <91951607+sofstam@users.noreply.github.com>
