### 2023-05-02 #84 @snacktavish
Hi @Midnighter  - It was the link from here: https://taxpasta.readthedocs.io/en/latest/tutorials/getting-started/#additional-functionality to here https://taxpasta.readthedocs.io/how-tos/how-to-add-names

The how to is there I think - just formatting issues.
