+++ docs/index.md
@@ -13,7 +13,7 @@ decided to take [the lessons learnt](https://xkcd.com/927/) to heart and provide
 our own solution to deal with this pasticcio. With taxpasta you can ingest all
 of those formats and, at a minimum, output taxonomy identifiers and their
 integer counts. Taxpasta can not only standardise profiles but also _merge_ them
-across samples for the same profiler into a single table.
+across samples for the _same_ profiler into a single table.
 
 ### Supported Taxonomic Profilers
 
