### 2023-05-06 bb303ff
docs: put how-tos into title case
