### 2023-05-06 79a00cf
docs: remove classifiers, use profilers
Add text and formatting.
