### 2023-05-06 d7a0a85
docs: remove classifiers, use profilers
Add text and formatting.
