### 2023-05-06 b71edfa
docs: add Python part to the merge how-to
