+++ docs/how-tos/how-to-merge-across-profilers.md
@@ -1,8 +1,8 @@
 # How-to Merge Across Profilers
 
-As stated in the main description of the tools and tutorials, `taxpasta`
-does not (directly) support merging _across_ different profilers, as each tool
-may have its own reference database, taxonomy, and/or abundance metric. This can
+As stated in the main description of the tools and tutorials, `taxpasta` does
+not (directly) support merging _across_ different profilers, as each tool may
+have its own reference database, taxonomy, and/or abundance metric. This can
 risk making naïve assumptions and false-positive interpretations, thus
 `taxpasta` is designed to help _prepare_ data for cross-profiler analysis
 without doing so itself. We highly recommend doing this mindfully in an
@@ -14,56 +14,106 @@ necessary.
 
 ## Dependencies
 
-You will need the following packages and libraries.
+=== "R"
 
-```r
-install.packages(
-    c("readr", "dplyr", "purrr", "tibble", "tidyr"),
-    dependencies = TRUE
-)
-```
+    You will need the following packages and libraries.
+
+    ```r
+    install.packages(
+        c("readr", "dplyr", "purrr", "tibble", "tidyr"),
+        dependencies = TRUE
+    )
+    ```
+
+=== "Python"
+
+    [Pandas](https://pandas.pydata.org/) is already part of the taxpasta
+    installation, so you don't need to install anything further.
 
 ## Merging Across Profilers
 
 Assuming you had two files in the same directory, `motus_dbMOTUs.tsv` and
-`kraken2_db2.tsv` - both of which are the output from a previous `taxpasta merge`
-command - you can load them as follows.
+`kraken2_db2.tsv` - both of which are the output from a previous `taxpasta merge` command - you can load them as follows.
 
-First, create a list of the TSV files.
+=== "R"
 
-```r
-filelist <- list.files(pattern = "*.tsv")
-```
+    First, create a list of the TSV files.
 
-Next, we use the list to create a table with the file names in one column, load the contents
-of each file into a nested column, and finally remove the nesting for those contents. If your
-files were created in wide format (the default for `taxpasta`), then we also need to pivot the
-tables to end up with a tidy format. Otherwise, the highlighted lines below won't be necessary[^1].
+    ```r
+    filelist <- list.files(pattern = "*.tsv")
+    ```
 
-```r hl_lines="6-11"
-filelist |>
-    as_tibble_col(column_name = "filename") |>
-    mutate(
-        file_contents = map(
-            filename, ~ read_tsv(.x) |>
-                pivot_longer(
-                    !matches("taxonomy_id"),
-                    names_to = "file",
-                    values_to = "value"
-                ) |>
-                select(taxonomy_id, file, value)
-        )
-    ) |>
-    select(file_contents) |>
-    unnest(cols = c(file_contents))
-```
+    Next, we use the list to create a table with the file names in one column,
+    load the contents of each file into a nested column, and finally remove the
+    nesting for those contents. If your files were created in wide format (the
+    default for `taxpasta`), then we also need to pivot the tables to end up
+    with a tidy format. Otherwise, the highlighted lines below won't be
+    necessary[^1].
 
-This will result in a _long_ format table containing three columns:
-`taxonomy_id`, `file`, and `value`.
+    ```r hl_lines="6-11" linenums="1"
+    filelist |>
+        as_tibble_col(column_name = "filename") |>
+        mutate(
+            file_contents = map(
+                filename, ~ read_tsv(.x) |>
+                    pivot_longer(  # (1)!
+                        !matches("taxonomy_id"),
+                        names_to = "sample",
+                        values_to = "count"
+                    ) |>
+                    select(taxonomy_id, sample, count)
+            )
+        ) |>
+        unnest(cols = file_contents)
+    ```
 
[…]
