### 2023-05-06 taxprofiler/taxpasta#100 @Midnighter
[BUG] Certain parser problems are only warnings
Certain parser issues should be raised as exceptions instead of warnings. Reported in https://github.com/pyOpenSci/software-submission/issues/84#issuecomment-1530191450.

```sh
taxpasta merge -p motus -o /tmp/motus.tsv \
    tests/data/motus/2612_pe-ERR5766176-db_mOTU.out \
    tests/data/motus/2612_se-ERR5766180-db_mOTU.out \
    tests/data/kraken2/2612_pe-ERR5766176-db1.kraken2.report.txt
```
