### 2023-05-06 b7966a6
feat: raise parser warning as exception
