### 2023-05-06 9addd69
feat: raise parser warning as exception
