+++ src/taxpasta/infrastructure/application/motus/motus_profile_reader.py
@@ -19,7 +19,10 @@
 """Provide a reader for motus profiles."""
 
 
+import warnings
+
 import pandas as pd
+from pandas.errors import ParserWarning
 from pandera.typing import DataFrame
 
 from taxpasta.application.service import BufferOrFilepath, ProfileReader
@@ -33,18 +36,27 @@ class MotusProfileReader(ProfileReader):
     @classmethod
     def read(cls, profile: BufferOrFilepath) -> DataFrame[MotusProfile]:
         """Read a mOTUs taxonomic profile from a file."""
-        result = pd.read_table(
-            filepath_or_buffer=profile,
-            sep="\t",
-            skiprows=3,
-            header=None,
-            names=[
-                MotusProfile.consensus_taxonomy,
-                MotusProfile.ncbi_tax_id,
-                MotusProfile.read_count,
-            ],
-            index_col=False,
-            dtype={MotusProfile.ncbi_tax_id: str},
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings(action="error", category=ParserWarning)
+            try:
+                result = pd.read_table(
+                    filepath_or_buffer=profile,
+                    sep="\t",
+                    skiprows=3,
+                    header=None,
+                    names=[
+                        MotusProfile.consensus_taxonomy,
+                        MotusProfile.ncbi_tax_id,
+                        MotusProfile.read_count,
+                    ],
+                    index_col=False,
+                    dtype={MotusProfile.ncbi_tax_id: str},
+                )
+            except ParserWarning as exc:
+                raise ValueError(
+                    "There were unexpected issues with the data. Please double-check "
+                    "that you chose a matching combination of metagenomic profiler and "
+                    "profile."
+                ) from exc
         cls._check_num_columns(result, MotusProfile)
         return result
