### 2023-05-08 8994269
chore: ignore pandas errors types
