+++ pyproject.toml
@@ -155,6 +155,7 @@ module = [
     "odf",
     "openpyxl",
     "pandas",
+    "pandas.errors",
     "pandera",
     "pandera.errors",
     "pandera.typing",
