### 2023-05-08 7f9a7e3
refactor: move warnings related code to decorator
