### 2023-05-08 e5680a3
refactor: stop coercion, test for empty
