+++ src/taxpasta/infrastructure/application/motus/motus_profile_reader.py
@@ -46,7 +46,7 @@ class MotusProfileReader(ProfileReader):
                 MotusProfile.read_count,
             ],
             index_col=False,
-            dtype={MotusProfile.ncbi_tax_id: str},
+            dtype={MotusProfile.ncbi_tax_id: "Int64"},
         )
         cls._check_num_columns(result, MotusProfile)
         return result
