### 2023-05-08 ee43ceb
refactor: stop coercion, test for empty
