+++ tests/integration/test_motus_etl.py
@@ -24,10 +24,30 @@ from pathlib import Path
 import pytest
 from pandera.errors import SchemaErrors
 
+from taxpasta.application.error import StandardisationError
 from taxpasta.infrastructure.application import (
     MotusProfileReader,
     MotusProfileStandardisationService,
 )
+
+
+@pytest.fixture(
+    scope="module",
+    params=[
+        ("bracken", "2612_pe-ERR5766176_B-db1_S.tsv"),
+        ("centrifuge", "AD_pe-db1.centrifuge.txt"),
+        ("diamond", "diamond_valid_1.tsv"),
+        ("kaiju", "barcode41_se-barcode41-kaiju.txt"),
+        ("kraken2", "2612_pe-ERR5766176-db1.kraken2.report.txt"),
+        ("krakenuniq", "test1.krakenuniq.report.txt"),
+        ("megan6", "malt_rma2info_valid.txt.gz"),
+        ("metaphlan", "mpa_valid_complex.tsv"),
+    ],
+)
+def other_profile(data_dir: Path, request: pytest.FixtureRequest) -> Path:
+    """Return paths to other profilers' profiles."""
+    base, filename = request.param
+    return data_dir / base / filename
 
 
 @pytest.mark.parametrize(
@@ -48,3 +68,11 @@ def test_read_correctness(
     MotusProfileStandardisationService.transform(
         MotusProfileReader.read(motus_data_dir / filename)
     )
+
+
+def test_failure_on_other_profiles(other_profile: Path):
+    """"""
+    with pytest.raises((TypeError, ValueError, SchemaErrors, StandardisationError)):
+        MotusProfileStandardisationService.transform(
+            MotusProfileReader.read(other_profile)
+        )
