### 2023-05-08 6e3dc43
refactor: introduce base model
