+++ src/taxpasta/infrastructure/application/motus/motus_profile.py
@@ -23,22 +23,12 @@ import pandas as pd
 import pandera as pa
 from pandera.typing import Series
 
+from taxpasta.infrastructure.helpers import BaseDataFrameModel
 
-class MotusProfile(pa.DataFrameModel):
+
+class MotusProfile(BaseDataFrameModel):
     """Define the expected mOTUs profile format."""
 
     consensus_taxonomy: Series[str] = pa.Field()
     ncbi_tax_id: Series[pd.Int64Dtype] = pa.Field(nullable=True)
     read_count: Series[int] = pa.Field(ge=0)
-
-    @pa.dataframe_check
-    def check_not_empty(cls, profile: pd.DataFrame) -> bool:
-        """Check that the read in profile is *not* empty."""
-        return not profile.empty
-
-    class Config:
-        """Configure the schema model."""
-
-        coerce = False
-        ordered = True
-        strict = True
