### 2023-05-08 11141ac
refactor: introduce base model
