### 2023-05-10 #84 @bluegenes
Hi folks, looks like you've addressed my concerns, so I'm happy to sign off.

> A bit of a tangent but since you're all experts @ctb, @bluegenes, and @snacktavish I wanted to ask your opinions on the following issue: Currently, taxpasta converts all relative abundances to integers. This design decision was based on my experience with some statistical tools requiring this. I think, it was vegan::adonis but I can't even remember properly. The other decent option is to convert everything to fractions. A third option is to provide a flag that lets users choose between either output. Do you have any opinions on this?

No strong opinions -- my personal preference is to not drop information (decimals) when possible, but I think it's fine to stick with integers if there are tools that require this.
