### 2023-05-27 db82318
refactor: adjust docstrings and variable naming
