+++ tests/unit/infrastructure/application/kraken2/test_kraken2_profile.py
@@ -29,61 +29,71 @@ from taxpasta.infrastructure.application import Kraken2Profile
 
 
 @pytest.mark.parametrize(
-    "columns",
+    "table",
     [
-        (
-            "percent",
-            "clade_assigned_reads",
-            "direct_assigned_reads",
-            "taxonomy_lvl",
-            "taxonomy_id",
-            "name",
+        pd.DataFrame(
+            {
+                "percent": pd.Series(data=[100.0], dtype=float),
+                "clade_assigned_reads": pd.Series(data=[100], dtype=int),
+                "direct_assigned_reads": pd.Series(data=[100], dtype=int),
+                "taxonomy_lvl": pd.Series(data=["R"], dtype=str),
+                "taxonomy_id": pd.Series(data=[1], dtype="Int64"),
+                "name": pd.Series(data=["root"], dtype=str),
+            }
         ),
-        (
-            "percent",
-            "clade_assigned_reads",
-            "direct_assigned_reads",
-            "num_minimizers",
-            "distinct_minimizers",
-            "taxonomy_lvl",
-            "taxonomy_id",
-            "name",
+        pd.DataFrame(
+            {
+                "percent": pd.Series(data=[100.0], dtype=float),
+                "clade_assigned_reads": pd.Series(data=[100], dtype=int),
+                "direct_assigned_reads": pd.Series(data=[100], dtype=int),
+                "num_minimizers": pd.Series(data=[1], dtype=int),
+                "distinct_minimizers": pd.Series(data=[1], dtype=int),
+                "taxonomy_lvl": pd.Series(data=["R"], dtype=str),
+                "taxonomy_id": pd.Series(data=[1], dtype="Int64"),
+                "name": pd.Series(data=["root"], dtype=str),
+            }
         ),
         pytest.param(
-            (
-                "percent",
-                "clade_assigned_reads",
-                "direct_assigned_reads",
-                "name",
+            pd.DataFrame(
+                {
+                    "percent": pd.Series(data=[100.0], dtype=float),
+                    "clade_assigned_reads": pd.Series(data=[100], dtype=int),
+                    "direct_assigned_reads": pd.Series(data=[100], dtype=int),
+                    "name": pd.Series(data=["root"], dtype=str),
+                }
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'taxonomy_lvl' not in dataframe"
             ),
         ),
         pytest.param(
-            (
-                "percent",
-                "clade_assigned_reads",
-                "direct_assigned_reads",
-                "num_minimizers",
-                "distinct_minimizers",
-                "taxonomy_lvl",
-                "taxonomy_id",
-                "name",
-                "rank",
+            pd.DataFrame(
+                {
+                    "percent": pd.Series(data=[100.0], dtype=float),
+                    "clade_assigned_reads": pd.Series(data=[100], dtype=int),
+                    "direct_assigned_reads": pd.Series(data=[100], dtype=int),
+                    "num_minimizers": pd.Series(data=[1], dtype=int),
+                    "distinct_minimizers": pd.Series(data=[1], dtype=int),
+                    "taxonomy_lvl": pd.Series(data=["R"], dtype=str),
+                    "taxonomy_id": pd.Series(data=[1], dtype="Int64"),
+                    "name": pd.Series(data=["root"], dtype=str),
+                    "rank": pd.Series(data=["R"], dtype=str),
+                }
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'rank' not in DataFrameSchema"
             ),
[…]
