### 2023-05-27 dcaed6c
refactor: inherit function attributes via @wraps
