### 2023-05-27 50b2904
refactor: make the Bracken validation more strict
