+++ tests/unit/infrastructure/application/bracken/test_bracken_profile.py
@@ -19,7 +19,7 @@
 """Test that the schema model validates bracken profiles correctly."""
 
 
-from typing import Collection
+from collections import OrderedDict
 
 import pandas as pd
 import pytest
@@ -29,54 +29,70 @@ from taxpasta.infrastructure.application import BrackenProfile
 
 
 @pytest.mark.parametrize(
-    "columns",
+    "profile",
     [
-        (
-            "name",
-            "taxonomy_id",
-            "taxonomy_lvl",
-            "kraken_assigned_reads",
-            "added_reads",
-            "new_est_reads",
-            "fraction_total_reads",
+        pd.DataFrame(
+            OrderedDict(
+                [
+                    ("name", ["root"]),
+                    ("taxonomy_id", [1]),
+                    ("taxonomy_lvl", ["root"]),
+                    ("kraken_assigned_reads", [100]),
+                    ("added_reads", [0]),
+                    ("new_est_reads", [100]),
+                    ("fraction_total_reads", [1.0]),
+                ]
+            )
         ),
         pytest.param(
-            (
-                "name",
-                "taxonomy_id",
-                "kraken_assigned_reads",
-                "added_reads",
-                "new_est_reads",
-                "fraction_total_reads",
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("name", ["root"]),
+                        ("taxonomy_id", [1]),
+                        ("kraken_assigned_reads", [100]),
+                        ("added_reads", [0]),
+                        ("new_est_reads", [100]),
+                        ("fraction_total_reads", [1.0]),
+                    ]
+                )
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'taxonomy_lvl' not in dataframe"
             ),
         ),
         pytest.param(
-            (
-                "name",
-                "taxonomy_id",
-                "taxonomy_lvl",
-                "kraken_assigned_reads",
-                "added_reads",
-                "new_est_reads",
-                "fraction_total_reads",
-                "foo",
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("name", ["root"]),
+                        ("taxonomy_id", [1]),
+                        ("taxonomy_lvl", ["root"]),
+                        ("kraken_assigned_reads", [100]),
+                        ("added_reads", [0]),
+                        ("new_est_reads", [100]),
+                        ("fraction_total_reads", [1.0]),
+                        ("foo", [1.0]),
+                    ]
+                )
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'foo' not in DataFrameSchema"
             ),
         ),
         pytest.param(
-            (
-                "name",
-                "taxonomy_lvl",
-                "added_reads",
-                "fraction_total_reads",
-                "taxonomy_id",
-                "new_est_reads",
-                "kraken_assigned_reads",
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("name", ["root"]),
+                        ("taxonomy_lvl", ["root"]),
+                        ("added_reads", [0]),
+                        ("kraken_assigned_reads", [100]),
+                        ("taxonomy_id", [1]),
+                        ("new_est_reads", [100]),
+                        ("fraction_total_reads", [1.0]),
+                    ]
+                )
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'taxonomy_lvl' out-of-order"
[…]
