### 2023-05-27 85bd6bd
refactor: make centrifuge validation more strict
