### 2023-05-27 c14a076
refactor: make centrifuge validation more strict
