### 2023-05-27 057669a
chore: add all profiles to fixture params
