+++ tests/integration/test_bracken_etl.py
@@ -38,6 +38,7 @@ from taxpasta.infrastructure.application import (
         ("centrifuge", "AD_pe-db1.centrifuge.txt"),
         ("diamond", "diamond_valid_1.tsv"),
         ("kaiju", "barcode41_se-barcode41-kaiju.txt"),
+        ("kraken2", "2612_pe-ERR5766176-db1.kraken2.report.txt"),
         ("krakenuniq", "test1.krakenuniq.report.txt"),
         ("megan6", "malt_rma2info_valid.txt.gz"),
         ("metaphlan", "mpa_valid_complex.tsv"),
