### 2023-05-27 757553c
chore: add all profiles to fixture params
