### 2023-05-27 2eb5c5b
refactor: make diamond validation more strict
