### 2023-05-27 5bb6dba
refactor: make diamond validation more strict
