### 2023-05-27 16ab6e6
refactor: make Kaiju validation more strict
