### 2023-05-27 67765c8
refactor: make MEGAN6 validation more strict
