### 2023-05-27 ac4e0d3
refactor: make MEGAN6 validation more strict
