### 2023-05-27 0511a54
fix: remove classmethod decorator
