+++ src/taxpasta/infrastructure/application/kaiju/kaiju_profile.py
@@ -37,14 +37,12 @@ class KaijuProfile(BaseDataFrameModel):
     taxon_name: Series[str] = pa.Field()
 
     @pa.check("percent", name="compositionality")
-    @classmethod
     def check_compositionality(cls, percent: Series[float]) -> bool:
         """Check that the percentages add up to a hundred."""
         # Kaiju reports percentages with sixth decimals
         return percent.empty or bool(np.isclose(percent.sum(), 100.0, atol=1.0))
 
     @pa.check("file", name="unique_filename")
-    @classmethod
     def check_unique_filename(cls, file_col: Series[str]) -> bool:
         """Check that Kaiju filename is unique."""
         return file_col.empty or file_col.nunique() == 1
