### 2023-05-27 2533eaf
fix: remove classmethod decorator
