### 2023-05-27 c6f57c3
tests: use ordered dicts
