+++ tests/unit/infrastructure/application/kraken2/test_kraken2_profile.py
@@ -19,7 +19,7 @@
 """Test that the schema model validates kraken2 profiles correctly."""
 
 
-from typing import Collection
+from collections import OrderedDict
 
 import pandas as pd
 import pytest
@@ -29,38 +29,45 @@ from taxpasta.infrastructure.application import Kraken2Profile
 
 
 @pytest.mark.parametrize(
-    "table",
+    "profile",
     [
         pd.DataFrame(
-            {
-                "percent": pd.Series(data=[100.0], dtype=float),
-                "clade_assigned_reads": pd.Series(data=[100], dtype=int),
-                "direct_assigned_reads": pd.Series(data=[100], dtype=int),
-                "taxonomy_lvl": pd.Series(data=["R"], dtype=str),
-                "taxonomy_id": pd.Series(data=[1], dtype="Int64"),
-                "name": pd.Series(data=["root"], dtype=str),
-            }
+            OrderedDict(
+                [
+                    ("percent", [100.0]),
+                    ("clade_assigned_reads", [100]),
+                    ("direct_assigned_reads", [100]),
+                    ("taxonomy_lvl", ["R"]),
+                    ("taxonomy_id", [1]),
+                    ("name", ["root"]),
+                ]
+            )
         ),
         pd.DataFrame(
-            {
-                "percent": pd.Series(data=[100.0], dtype=float),
-                "clade_assigned_reads": pd.Series(data=[100], dtype=int),
-                "direct_assigned_reads": pd.Series(data=[100], dtype=int),
-                "num_minimizers": pd.Series(data=[1], dtype=int),
-                "distinct_minimizers": pd.Series(data=[1], dtype=int),
-                "taxonomy_lvl": pd.Series(data=["R"], dtype=str),
-                "taxonomy_id": pd.Series(data=[1], dtype="Int64"),
-                "name": pd.Series(data=["root"], dtype=str),
-            }
+            OrderedDict(
+                [
+                    ("percent", [100.0]),
+                    ("clade_assigned_reads", [100]),
+                    ("direct_assigned_reads", [100]),
+                    ("num_minimizers", [1]),
+                    ("distinct_minimizers", [1]),
+                    ("taxonomy_lvl", ["R"]),
+                    ("taxonomy_id", [1]),
+                    ("name", ["root"]),
+                ]
+            )
         ),
         pytest.param(
             pd.DataFrame(
-                {
-                    "percent": pd.Series(data=[100.0], dtype=float),
-                    "clade_assigned_reads": pd.Series(data=[100], dtype=int),
-                    "direct_assigned_reads": pd.Series(data=[100], dtype=int),
-                    "name": pd.Series(data=["root"], dtype=str),
-                }
+                OrderedDict(
+                    [
+                        ("percent", [100.0]),
+                        ("clade_assigned_reads", [100]),
+                        ("direct_assigned_reads", [100]),
+                        ("taxonomy_id", [1]),
+                        ("name", ["root"]),
+                    ]
+                )
             ),
             marks=pytest.mark.raises(
                 exception=SchemaError, message="column 'taxonomy_lvl' not in dataframe"
@@ -68,17 +75,19 @@ from taxpasta.infrastructure.application import Kraken2Profile
         ),
         pytest.param(
             pd.DataFrame(
-                {
-                    "percent": pd.Series(data=[100.0], dtype=float),
-                    "clade_assigned_reads": pd.Series(data=[100], dtype=int),
-                    "direct_assigned_reads": pd.Series(data=[100], dtype=int),
-                    "num_minimizers": pd.Series(data=[1], dtype=int),
-                    "distinct_minimizers": pd.Series(data=[1], dtype=int),
[…]
