### 2023-05-27 f1edfea
tests: use ordered dicts
