+++ tests/integration/test_kaiju_etl.py
@@ -37,7 +37,7 @@ from taxpasta.infrastructure.application import (
         ("bracken", "2612_pe-ERR5766176_B-db1_S.tsv"),
         ("centrifuge", "AD_pe-db1.centrifuge.txt"),
         ("diamond", "diamond_valid_1.tsv"),
-        # ("kaiju", "barcode41_se-barcode41-kaiju.txt"),
+        # ("kaiju", "barcode41_se-barcode41-kaiju.txt"),  # noqa: E800
         ("kraken2", "2612_pe-ERR5766176-db1.kraken2.report.txt"),
         ("krakenuniq", "test1.krakenuniq.report.txt"),
         ("megan6", "malt_rma2info_valid.txt.gz"),
