+++ tests/integration/test_metaphlan_etl.py
@@ -41,7 +41,7 @@ from taxpasta.infrastructure.application import (
         ("kraken2", "2612_pe-ERR5766176-db1.kraken2.report.txt"),
         ("krakenuniq", "test1.krakenuniq.report.txt"),
         ("megan6", "malt_rma2info_valid.txt.gz"),
-        # ("metaphlan", "mpa_valid_complex.tsv"),
+        # ("metaphlan", "mpa_valid_complex.tsv"),  # noqa: E800
         ("motus", "2612_pe-ERR5766176-db_mOTU.out"),
     ],
 )
