+++ docs/commands/index.md
@@ -4,5 +4,5 @@ Taxpasta offers multiple commands that perform different functionality.
 
 The following pages document the 'what, how, and why' for each command.
 
--   [`taxpasta standardise`](standardise)
--   [`taxpasta merge`](merge)
+-   [`taxpasta standardise`](standardise.md)
+-   [`taxpasta merge`](merge.md)
