+++ docs/quick_reference/index.md
@@ -2,5 +2,5 @@
 
 The following pages provide the `help` messages of each command to act as a quick reference for how to use them.
 
--   [`taxpasta standardise`](standardise)
--   [`taxpasta merge`](merge)
+-   [`taxpasta standardise`](standardise.md)
+-   [`taxpasta merge`](merge.md)
