### 2023-05-28 90df6ce
refactor: change error message
