### 2023-05-28 #84 @Midnighter
Dear @snacktavish,

Thank you again for your review. I will respond below to your points (quoted).

> In the tutorial, a little more biological context pn the meaning of the results would be useful.

> E.g. in the step head 2612_pe-ERR5766176-db1_kraken2.tsv

> These results look a bit concerning to me! What are soemtings things in taxon 0? Are all of the results not id'd to taxon? (This may be a limitation of this my own knowledge of taxonomic profiling, but including some biological context with the technical info can be helpful for users.)

From my point of view, providing interpretation or meaning is down to the individual profilers. Taxpasta is a pure utility and I would assume that its only users are those who have already chosen to run a particular profiler on their data. (For the specific example, taxon 0 is indeed unclassified. That's not surprising since it comes from a very small sequencing file that we run for testing purposes only.)

> I tried going off-script in the tutorial to see if I could merge the Kraken file with the motus files (this may not make a lot of sense to do, biologically, but I was curious) using:
taxpasta merge -p motus -o dbMOTUs_motus.tsv 2612_pe-ERR5766176-db_mOTU.out 2612_se-ERR5766180-db_mOTU.out 2612_pe-ERR5766176-db1.kraken2.report.txt

> It appeared to work... but the results were odd. At the end it showed a taxonomy id of 152 with 152 samples in 2612_pe-ERR5766176-db1.kraken2.report, whereas in the original table it looks like it should be taxon id 9606 with 152 samples.
Looking at the docs I think the software is only intended to merge profiles generated by the same tool, but the way that the description is set up, I thought merging profiles across tools was a key advantage of the software. (This is user error - but user error is one thing you can always rely on!!)
This looks like a bug? or maybe the software should have thrown an error instead of running? But users will definitely try to run things that are not quite right, and the software should do some checks on input file format rather than chugging along and outputting seemingly sensible but incorrect outputs.
(This may connect to the "Danger message" https://taxpasta.readthedocs.io/en/latest/commands/merge/#why here, but it does not even appear to be merging on identifier...)

Thank you for doing that. Your experience here prompted us to do some [fairly major refactoring](https://github.com/taxprofiler/taxpasta/pull/101), which made the whole profile validation part much more strict. We have thus ensured that providing an input from another than the chosen profiler will always result in an error. (Except for centrifuge and kraken2 which use the same six column layout.) We have also added tests to verify that errors occur for all profilers. Due to that work our response is also a bit later than hoped.

> I wanted to add taxon names to try to make the outputs make more sense to me, but I wasn't able to read the documentation on that at
https://taxpasta.readthedocs.io/how-tos/how-to-add-names

Thank you for noticing this. We have [fixed links in the documentation](https://github.com/taxprofiler/taxpasta/pull/103).

> The error messages in other places could also use some work - I ran:
taxpasta merge -p motus -o dbMOTUs_motus.tsv 2612_pe-ERR5766176-db_mOTU.out 2612_se-ERR5766180-db_mOTU.out 2612_pe-ERR5766176-db1_kraken2.tsv
Which I think is just not a reasonable file format to try to merge, but the error did not really convey that (although it did point to the problematic file).

This is the same point as above, as far as I can tell.

> I was not able to find the paper.md file to review for JOSS, but I see that you are planning on that later.

We hope that we addressed your major points and will begin work on this paper. Please note that if you wanted to test the changes, you will have to install the `dev` branch, since we have not released these changes yet.
