### 2023-05-29 a6facb7
chore: prepare project setup for paper draft
