+++ .github/workflows/draft-pdf.yml
@@ -15,7 +15,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Draft PDF
-        uses: openjournals/openjournals-draft-action@v1
+        uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
           paper-path: paper/paper.md
