### 2023-05-29 ec97fad
chore: fix markdown draft command
The PDF needs to depend on the markdown, such that changes are tracked.
