+++ paper/paper.md
@@ -1,0 +1,38 @@
+---
+title: "TAXPASTA: TAXonomic Profile Aggregation and STAndardisation"
+tags:
+    - bioinformatics
+    - metagenomics
+    - profiling
+    - classification
+    - standardisation
+    - taxonomy
+    - Python
+authors:
+    - name: Moritz E. Beber
+      orcid: 0000-0003-2406-1978
+      corresponding: true
+      affiliation: 1
+    - name: Maxime Borry
+      orcid: 0000-0001-9140-7559
+      affiliation: 2
+    - name: Sofia Stamouli
+      orcid: 0009-0006-0893-3771
+      affiliation: 3
+    - name: James A. Fellows Yates
+      orcid: 0000-0001-5585-6277
+      affiliation: "4, 2"
+affiliations:
+    - name: Unseen Bio ApS, Fruebjergvej 3, 2100 Copenhagen,Capital Region, Denmark
+      index: 1
+    - name: Department of Archaeogenetics, Max Planck Institute for Evolutionary Anthropology, Deutscher Platz 6, 4103 Leipzig, Saxony, Germany
+      index: 2
+    - name: Department of Microbiology, Tumor and Cell Biology, Karolinska Institute, Solnavägen 1, 171 77 Solna, Stockholm, Sweden
+      index: 3
+    - name: Department of Paleobiotechnology, Leibniz Institute for Natural Product Research and Infection Biology Hans Knöll Institute, Adolf-Reichwein-Straße 23, 7745 Jena, Thuringia, Germany
+      index: 4
+date: 2023-05-29
+bibliography: paper.bib
+---
+
+# References
