+++ paper/paper.md
@@ -23,16 +23,93 @@ authors:
       orcid: 0000-0001-5585-6277
       affiliation: "4, 2"
 affiliations:
-    - name: Unseen Bio ApS, Fruebjergvej 3, 2100 Copenhagen,Capital Region, Denmark
+    - name: Unseen Bio ApS, Fruebjergvej 3, 2100 Copenhagen, Denmark
       index: 1
-    - name: Department of Archaeogenetics, Max Planck Institute for Evolutionary Anthropology, Deutscher Platz 6, 4103 Leipzig, Saxony, Germany
+    - name: Department of Archaeogenetics, Max Planck Institute for Evolutionary Anthropology, Deutscher Platz 6, 04103 Leipzig, Germany
       index: 2
-    - name: Department of Microbiology, Tumor and Cell Biology, Karolinska Institute, Solnavägen 1, 171 77 Solna, Stockholm, Sweden
+    - name: Department of Microbiology, Tumor and Cell Biology, Karolinska Institute, Solnavägen 1, 171 77 Solna, Sweden
       index: 3
-    - name: Department of Paleobiotechnology, Leibniz Institute for Natural Product Research and Infection Biology Hans Knöll Institute, Adolf-Reichwein-Straße 23, 7745 Jena, Thuringia, Germany
+    - name: Department of Paleobiotechnology, Leibniz Institute for Natural Product Research and Infection Biology Hans Knöll Institute, Adolf-Reichwein-Straße 23, 07745 Jena, Germany
       index: 4
 date: 2023-05-29
 bibliography: paper.bib
 ---
+
+# Summary
+
+Metagenomic analysis is a scientific field that was strongly enabled by
+next-generation sequencing. It is chiefly concerned with the taxonomic and
+functional composition of microbial communities in their natural habitats. That
+means, researchers would like to know from metagenomic sequencing samples who
+(what microbes) is present, and what are they doing (what functions are they
+performing)? The nature of this field is such that it intersects with ecology,
+statistics, and bioinformatics.
+
+In part, due to the interdisciplinary nature of the field, there exists a
+diverse number of bioinformatics tools in order to analyse metagenomic
+sequencing data and produce metagenomic profiles; answering to the question of
+who is there and in what relative proportion. Most of those tools have invented
+their own tabular result format, which complicates downstream analysis and in
+particular comparison across tools.
+
+# Statement of need
+
+TAXPASTA is a Python package for standardising and aggregating metagenomic
+profiles coming from a wide range of tools and databases. It was developed as
+part of the nf-core/taxprofiler pipeline [@nf-core/taxprofiler:2023] within the
+nf-core community [@ewels_nf-core_2020].
+
+Across profilers, relative abundances can be reported in read counts, fractions,
+or percentages, as well as any number of additional columns with extra
+information. TAXPASTA can ingest all of those formats and, at a minimum, produce
+a standardised output containing taxonomic identifiers and their relative
+abundances as integer counts. It can also be used to aggregate profiles from
+the same profiler and merge them into a single, standardised table. Having a
+singular format facilitates downstream analyses and comparisons.
+
+Primarily, TAXPASTA is a command-line tool that is designed to be used as a
+building block in metagenomic analysis workflows. At the time of writing, it was
+able to read profiles from nine different profilers, namely Bracken
+[@lu_bracken_2017], Centrifuge [@kim_centrifuge_2016], DIAMOND
+[@buchfink_sensitive_2021], Kaiju [@menzel_fast_2016], Kraken2
[…]
