+++ paper/paper.bib
@@ -1,0 +1,189 @@
+
+@article{lu_bracken_2017,
+	title = {Bracken: estimating species abundance in metagenomics data},
+	volume = {3},
+	issn = {2376-5992},
+	url = {https://doi.org/10.7717/peerj-cs.104},
+	doi = {10.7717/peerj-cs.104},
+	abstract = {Metagenomic experiments attempt to characterize microbial communities using high-throughput DNA sequencing. Identification of the microorganisms in a sample provides information about the genetic profile, population structure, and role of microorganisms within an environment. Until recently, most metagenomics studies focused on high-level characterization at the level of phyla, or alternatively sequenced the 16S ribosomal RNA gene that is present in bacterial species. As the cost of sequencing has fallen, though, metagenomics experiments have increasingly used unbiased shotgun sequencing to capture all the organisms in a sample. This approach requires a method for estimating abundance directly from the raw read data. Here we describe a fast, accurate new method that computes the abundance at the species level using the reads collected in a metagenomics experiment. Bracken (Bayesian Reestimation of Abundance after Classification with KrakEN) uses the taxonomic assignments made by Kraken, a very fast read-level classifier, along with information about the genomes themselves to estimate abundance at the species level, the genus level, or above. We demonstrate that Bracken can produce accurate species- and genus-level abundance estimates even when a sample contains multiple near-identical species.},
+	journal = {PeerJ Computer Science},
+	author = {Lu, Jennifer and Breitwieser, Florian P. and Thielen, Peter and Salzberg, Steven L.},
+	editor = {Procter, James},
+	month = jan,
+	year = {2017},
+	keywords = {Bayesian estimation, Metagenomics, Microbiome, Species abundance},
+	pages = {e104},
+	file = {Lu et al (2017)_PeerJ Computer Science.pdf:/home/moritz/Dropbox/References/Storage/Lu et al (2017)_PeerJ Computer Science.pdf:application/pdf},
+}
+
+@article{wood_improved_2019,
+	title = {Improved metagenomic analysis with {Kraken} 2},
+	volume = {20},
+	issn = {1474-760X},
+	url = {https://doi.org/10.1186/s13059-019-1891-0},
+	doi = {10.1186/s13059-019-1891-0},
+	abstract = {Although Kraken’s k-mer-based approach provides a fast taxonomic classification of metagenomic sequence data, its large memory requirements can be limiting for some applications. Kraken 2 improves upon Kraken 1 by reducing memory usage by 85\%, allowing greater amounts of reference genomic data to be used, while maintaining high accuracy and increasing speed fivefold. Kraken 2 also introduces a translated search mode, providing increased sensitivity in viral metagenomics analysis.},
+	number = {1},
+	journal = {Genome Biology},
+	author = {Wood, Derrick E. and Lu, Jennifer and Langmead, Ben},
+	month = nov,
+	year = {2019},
+	pages = {257},
+	file = {Wood et al (2019)_Genome Biology.pdf:/home/moritz/Dropbox/References/Storage/Wood et al (2019)_Genome Biology.pdf:application/pdf},
+}
+
+@article{ruscheweyh_cultivation-independent_2022,
+	title = {Cultivation-independent genomes greatly expand taxonomic-profiling capabilities of {mOTUs} across various environments},
+	volume = {10},
+	issn = {2049-2618},
+	url = {https://doi.org/10.1186/s40168-022-01410-z},
+	doi = {10.1186/s40168-022-01410-z},
[…]
