+++ paper/paper.md
@@ -15,54 +15,42 @@ authors:
       affiliation: 1
     - name: Maxime Borry
       orcid: 0000-0001-9140-7559
-      affiliation: 2,3
+      affiliation: 2
     - name: Sofia Stamouli
       orcid: 0009-0006-0893-3771
-      affiliation: 4
+      affiliation: 3
     - name: James A. Fellows Yates
       orcid: 0000-0001-5585-6277
-      affiliation: "2, 3, 5"
+      affiliation: "4, 2"
 affiliations:
     - name: Unseen Bio ApS, Fruebjergvej 3, 2100 Copenhagen, Denmark
       index: 1
-    - name: Microbiome Sciences Group, Department of Archaeogenetics, Max Planck Institute for Evolutionary Anthropology, Deutscher Platz 6, 04103 Leipzig, Germany
+    - name: Department of Archaeogenetics, Max Planck Institute for Evolutionary Anthropology, Deutscher Platz 6, 04103 Leipzig, Germany
       index: 2
-    - name: Associated Research Group of Archaeogenetics, Leibniz Institute for Natural Product Research and Infection Biology Hans Knöll Institute, Adolf-Reichwein-Straße 23, 07745 Jena, Germany
-      index: 3
     - name: Department of Microbiology, Tumor and Cell Biology, Karolinska Institute, Solnavägen 1, 171 77 Solna, Sweden
-      index: 4
+      index: 3
     - name: Department of Paleobiotechnology, Leibniz Institute for Natural Product Research and Infection Biology Hans Knöll Institute, Adolf-Reichwein-Straße 23, 07745 Jena, Germany
-      index: 5
-
+      index: 4
 date: 2023-05-29
 bibliography: paper.bib
 ---
 
 # Summary
 
-Metagenomic analysis is chiefly concerned with untargeted genetic
-characterisation of the taxonomic and functional composition of whole
-communities of organisms. Researchers ask questions from metagenomic sequencing
-such as 'who is present' (what organisms), and 'what are they doing' (what
-functions are they performing)? The nature of this field is such that it
-intersects with ecology, medicine, statistics, and bioinformatics. Facilitated
-by the development of Next-Generation-Sequencing (NGS), the field often
-generates large datasets of many samples and many DNA sequences.
+Metagenomic analysis is a scientific field that was strongly enabled by
+next-generation sequencing. It is chiefly concerned with the taxonomic and
+functional composition of microbial communities in their natural habitats. That
+means, researchers would like to know from metagenomic sequencing samples who
+(what microbes) is present, and what are they doing (what functions are they
+performing)? The nature of this field is such that it intersects with ecology,
+statistics, and bioinformatics.
 
-In part, due to the interdisciplinary nature of the field, but also due to the
-size of datasets, the task of efficiently identifying the taxonomic 'origin' of
-each DNA molecule is a popular and unresolved bioinformatic problem. Thus, there
-exists a diverse number of bioinformatics tools in order to analyse metagenomic
-sequencing data and produce metagenomic profiles, and researchers will often.
-However, most of those tools have invented their own tabular result format,
-which complicates downstream analysis and in particular comparison across tools.
-
-TAXPASTA is a standalone command-line tool written in Python that aims to
-standardise the diverse range of metagenomic profiler output formats to simple
-tabular formats that are easy to use in downstream applications. TAXPASTA
-facilitates easier cross-comparison between taxonomic profiling tools without
-the need for external or dedicated modules or plugins needed of other
-'dedicated' metagenomic profile formats.
[…]
