+++ paper/paper.bib
@@ -156,21 +156,44 @@
 	file = {Breitwieser et al (2018)_Genome Biology.pdf:/home/moritz/Dropbox/References/Storage/Breitwieser et al (2018)_Genome Biology.pdf:application/pdf},
 }
 
-@article{gautam_meganserver_2023,
-	title = {{MeganServer}: facilitating interactive access to metagenomic data on a server},
-	volume = {39},
-	issn = {1367-4811},
-	url = {https://doi.org/10.1093/bioinformatics/btad105},
-	doi = {10.1093/bioinformatics/btad105},
-	abstract = {Metagenomic projects often involve large numbers of large sequencing datasets (totaling hundreds of gigabytes of data). Thus, computational preprocessing and analysis are usually performed on a server. The results of such analyses are then usually explored interactively. One approach is to use MEGAN, an interactive program that allows analysis and comparison of metagenomic datasets. Previous releases have required that the user first download the computed data from the server, an increasingly time-consuming process. Here, we present MeganServer, a stand-alone program that serves MEGAN files to the web, using a RESTful API, facilitating interactive analysis in MEGAN, without requiring prior download of the data. We describe a number of different application scenarios.MeganServer is provided as a stand-alone program tools/megan-server in the MEGAN software suite, available at https://software-ab.cs.uni-tuebingen.de/download/megan6. Source is available at: https://github.com/husonlab/megan-ce/tree/master/src/megan/ms.Supplementary data are available at Bioinformatics online.},
-	number = {3},
-	urldate = {2023-05-29},
-	journal = {Bioinformatics},
-	author = {Gautam, Anupam and Zeng, Wenhuan and Huson, Daniel H},
-	month = mar,
-	year = {2023},
-	pages = {btad105},
-	file = {Gautam et al (2023)_Bioinformatics.pdf:/home/moritz/Dropbox/References/Storage/Gautam et al (2023)_Bioinformatics.pdf:application/pdf},
+@ARTICLE{huson_megan_2016,
+  title    = "{MEGAN} Community Edition - Interactive Exploration and Analysis
+              of Large-Scale Microbiome Sequencing Data",
+  author   = "Huson, Daniel H and Beier, Sina and Flade, Isabell and Górska,
+              Anna and El-Hadidi, Mohamed and Mitra, Suparna and Ruscheweyh,
+              Hans-Joachim and Tappu, Rewati",
+  journal  = "PLoS computational biology",
+  volume   =  12,
+  number   =  6,
+  pages    = "e1004957",
+  month    =  jun,
+  year     =  2016,
+  url      = "http://dx.doi.org/10.1371/journal.pcbi.1004957",
+  doi      = "10.1371/journal.pcbi.1004957",
+  pmc      = "PMC4915700",
+  pmid     =  27327495,
+  issn     = "1553-734X,1553-7358",
+  language = "en"
+}
+
+@ARTICLE{vagene_malt_2018,
+  title    = "Salmonella enterica genomes from victims of a major
+              sixteenth-century epidemic in Mexico",
+  author   = "Vågene, Åshild J and Herbig, Alexander and Campana, Michael G and
+              Robles García, Nelly M and Warinner, Christina and Sabin, Susanna
+              and Spyrou, Maria A and Andrades Valtueña, Aida and Huson, Daniel
+              and Tuross, Noreen and Bos, Kirsten I and Krause, Johannes",
+  journal  = "Nature ecology \& evolution",
+  volume   =  2,
+  number   =  3,
+  pages    = "520--528",
+  month    =  jan,
+  year     =  2018,
+  url      = "http://dx.doi.org/10.1038/s41559-017-0446-6",
[…]
