### 2023-05-31 70266ce
Revert "Update affilaitions, acknowledgements, and funding, and commetns on summary"
This reverts commit 6f0f44ec69f3698bf1a3a8b06707e6727805fb36.
