+++ src/taxpasta/infrastructure/application/__init__.py
@@ -61,6 +61,11 @@ from .motus import (
     MotusProfileReader,
     MotusProfileStandardisationService,
 )
+from .ganon import (
+    GanonProfile,
+    GanonProfileReader,
+    GanonProfileStandardisationService,
+)
 from .sample_etl_application import SampleETLApplication
 from .standard_profile_file_format import StandardProfileFileFormat
 from .table_reader_file_format import TableReaderFileFormat

[…]
