+++ tests/unit/infrastructure/application/ganon/test_ganon_profile.py
@@ -1,0 +1,283 @@
+# Copyright (c) 2023 Moritz E. Beber
+# Copyright (c) 2023 Maxime Borry
+# Copyright (c) 2023 James A. Fellows Yates
+# Copyright (c) 2023 Sofia Stamouli.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+"""Test that the schema model validates ganon profiles correctly."""
+
+
+from collections import OrderedDict
+
+import pandas as pd
+import pytest
+from pandera.errors import SchemaError
+
+from taxpasta.infrastructure.application import GanonProfile
+
+
+@pytest.mark.parametrize(
+    "profile",
+    [
+        pd.DataFrame(
+            OrderedDict(
+                [
+                    ("rank", ["root"]),
+                    ("target", [1]),
+                    ("lineage", ["1"]),
+                    ("name", ["root"]),
+                    ("nr_unique", [1]),
+                    ("nr_shared", [1]),
+                    ("nr_children", [1]),
+                    ("nr_cumulative", [1]),
+                    ("pc_cumulative", [100.0]),
+                ]
+            )
+        ),
+        pytest.param(
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("rank", ["root"]),
+                        ("target", [1]),
+                        ("lineage", ["1"]),
+                        ("name", ["root"]),
+                        ("nr_unique", [1]),
+                        ("nr_shared", [1]),
+                        ("nr_children", [1]),
+                        ("nr_cumulative", [1]),
+                    ]
+                )
+            ),
+            marks=pytest.mark.raises(
+                exception=SchemaError, message="column 'pc_cumulative' not in dataframe"
+            ),
+        ),
+        pytest.param(
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("target", [1]),
+                        ("lineage", ["1"]),
+                        ("name", ["root"]),
+                        ("nr_unique", [1]),
+                        ("nr_shared", [1]),
+                        ("nr_children", [1]),
+                        ("nr_cumulative", [1]),
+                        ("pc_cumulative", [100.0]),
+                        ("rank", ["root"]),
+                    ]
+                )
+            ),
+            marks=pytest.mark.raises(
+                exception=SchemaError, message="column 'rank' out-of-order"
+            ),
+        ),
+    ],
+)
+def test_column_presence(profile: pd.DataFrame):
+    """Test that column names and order are validated."""
+    GanonProfile.validate(profile)
+
+
+@pytest.mark.parametrize(
+    "profile",
+    [
+        pd.DataFrame(
+            OrderedDict(
+                [
+                    ("rank", ["root"]),
+                    ("target", [1]),
+                    ("lineage", ["1"]),
+                    ("name", ["root"]),
+                    ("nr_unique", [1]),
+                    ("nr_shared", [1]),
+                    ("nr_children", [1]),
+                    ("nr_cumulative", [1]),
+                    ("pc_cumulative", [100.0]),
+                ]
+            )
+        ),
+        pytest.param(
+            pd.DataFrame(
+                OrderedDict(
+                    [
+                        ("rank", ["unclassified", "root"]),
+                        ("target", ["-", 1]),
[…]
