### 2023-06-24 83c7cb2
Apply suggestions from code review
Co-authored-by: Moritz E. Beber <midnighter@posteo.net>
