### 2023-06-24 82b23f0
Apply suggestions from code review
Co-authored-by: Moritz E. Beber <midnighter@posteo.net>
