+++ src/taxpasta/infrastructure/application/supported_profiler.py
@@ -29,10 +29,10 @@ class SupportedProfiler(str, Enum):
     bracken = "bracken"
     centrifuge = "centrifuge"
     diamond = "diamond"
+    ganon = "ganon"
     kaiju = "kaiju"
     kraken2 = "kraken2"
     krakenuniq = "krakenuniq"
     megan6 = "megan6"
     metaphlan = "metaphlan"
     motus = "motus"
-    ganon = "ganon"
