### 2023-06-24 43c3a82
refactor: further refinement of the paper
* Reformulate certain sentences
* Add paragraph on OPAL
* Provide further references and links
