+++ paper/paper.bib
@@ -1,4 +1,20 @@
 
+@article{huson_megan_2016,
+	title = {{MEGAN} {Community} {Edition} - {Interactive} {Exploration} and {Analysis} of {Large}-{Scale} {Microbiome} {Sequencing} {Data}},
+	volume = {12},
+	url = {https://doi.org/10.1371/journal.pcbi.1004957},
+	doi = {10.1371/journal.pcbi.1004957},
+	abstract = {Author Summary Microbiome sequencing projects continue to grow rapidly, both in the number of samples considered and sequencing reads collected. With MEGAN Community Edition (CE), we provide a highly efficient program for interactive analysis and comparison of such data, allowing one to explore hundreds of samples and billions of reads. While taxonomic profiling is performed based on the NCBI taxonomy, we provide a number of different functional profiling approaches such as SEED, eggNOG, KEGG, and a new InterPro2GO classification scheme. MEGAN CE also supports the use of metadata in the context of principal coordinate analysis and clustering analysis.},
+	number = {6},
+	journal = {PLOS Computational Biology},
+	author = {Huson, Daniel H. and Beier, Sina and Flade, Isabell and Górska, Anna and El-Hadidi, Mohamed and Mitra, Suparna and Ruscheweyh, Hans-Joachim and Tappu, Rewati},
+	month = jun,
+	year = {2016},
+	note = {Publisher: Public Library of Science},
+	pages = {e1004957},
+	file = {Huson et al (2016)_PLOS Computational Biology.PDF:/home/moritz/Dropbox/References/Storage/Huson et al (2016)_PLOS Computational Biology.PDF:application/pdf},
+}
+
 @article{lu_bracken_2017,
 	title = {Bracken: estimating species abundance in metagenomics data},
 	volume = {3},
@@ -11,7 +27,7 @@
 	editor = {Procter, James},
 	month = jan,
 	year = {2017},
-	keywords = {Bayesian estimation, Metagenomics, Microbiome, Species abundance},
+	keywords = {Metagenomics, Microbiome, Bayesian estimation, Species abundance},
 	pages = {e104},
 	file = {Lu et al (2017)_PeerJ Computer Science.pdf:/home/moritz/Dropbox/References/Storage/Lu et al (2017)_PeerJ Computer Science.pdf:application/pdf},
 }
@@ -32,6 +48,38 @@
 	file = {Wood et al (2019)_Genome Biology.pdf:/home/moritz/Dropbox/References/Storage/Wood et al (2019)_Genome Biology.pdf:application/pdf},
 }
 
+@article{sczyrba_critical_2017,
+	title = {Critical {Assessment} of {Metagenome} {Interpretation}—a benchmark of metagenomics software},
+	volume = {14},
+	issn = {1548-7105},
+	url = {https://doi.org/10.1038/nmeth.4458},
+	doi = {10.1038/nmeth.4458},
+	abstract = {The Critical Assessment of Metagenome Interpretation (CAMI) community initiative presents results from its first challenge, a rigorous benchmarking of software for metagenome assembly, binning and taxonomic profiling.},
+	number = {11},
+	journal = {Nature Methods},
[…]
