+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
       affiliation: 1
     - name: Maxime Borry
       orcid: 0000-0001-9140-7559
-      affiliation: 2,3
+      affiliation: "2, 3"
     - name: Sofia Stamouli
       orcid: 0009-0006-0893-3771
       affiliation: 4
@@ -47,22 +47,26 @@ such as 'who is present' (what organisms), and 'what are they doing' (what
 functions are they performing)? The nature of this field is such that it
 intersects with ecology, medicine, statistics, and bioinformatics. Facilitated
 by the development of Next-Generation-Sequencing (NGS), the field often
-generates large datasets of many samples and many DNA sequences.
+generates large datasets consisting of many samples (hundreds) and many
+sequencing reads (tens of millions).
 
-In part, due to the interdisciplinary nature of the field, but also due to the
-size of datasets, the task of efficiently identifying the taxonomic 'origin' of
-each DNA molecule is a popular and unresolved bioinformatic problem. Thus, there
-exists a diverse number of bioinformatics tools in order to analyse metagenomic
-sequencing data and produce metagenomic profiles, and researchers will often.
-However, most of those tools have invented their own tabular result format,
-which complicates downstream analysis and in particular comparison across tools.
+In part, due to the interdisciplinary nature of the field, but more importantly,
+due to the lack of a gold standard, the task of accurately identifying the
+taxonomic origin of each sequencing read is a popular and unresolved
+bioinformatics problem. Furthermore, the sizes of the datasets present
+interesting challenges for computational efficiency, which may require trading
+off accuracy for speed and memory use. Thus, there exists a diverse number of
+bioinformatics tools in order to analyse metagenomic sequencing data and produce
+metagenomic profiles. However, most of those tools have invented their own
+(often tabular) result formats, which complicates downstream analysis and in
+particular comparison across tools.
 
 TAXPASTA is a standalone command-line tool written in Python that aims to
 standardise the diverse range of metagenomic profiler output formats to simple
-tabular formats that are easy to use in downstream applications. TAXPASTA
-facilitates easier cross-comparison between taxonomic profiling tools without
-the need for external or dedicated modules or plugins needed of other
-'dedicated' metagenomic profile formats.
+tabular formats that are readily consumed in downstream applications. TAXPASTA
+facilitates cross-comparison between taxonomic profiling tools without the need
+for external or dedicated modules or plugins needed of other 'dedicated'
+metagenomic profile formats.
 
 # Statement of need
 
@@ -73,71 +77,89 @@ nf-core community [@ewels_nf-core_2020].
 
 Across profilers, relative abundances can be reported in read counts, fractions,
 or percentages, as well as any number of additional columns with extra
-information. Taxonomic identifiers can be recorded with taxonmic IDs, taxonomic
-names and/or in some cases semi-colon separated taxonomic 'paths'. These can
-also be formatted in different ways, from typical tables, to including
-'indented' taxonomy trees such as in the Kraken [@wood_improved_2019] family of
-profilers. Manually parsing these for comparison can be arduous task, with
-researchers often reverting to custom R, Python scripting, or even manual
-correction in spreadsheet software.
+information. Taxonomic identifiers can be recorded with taxonmic identifiers,
+taxonomic names and/or in some cases semi-colon-separated taxonomic 'paths'
+(lineages). These can also be formatted in different ways, from typical tables,
[…]
