### 2023-06-24 e3b0d5b
refactor: make compositionality check more tolerant
