### 2023-06-24 1f9afaa
fix: add missing pytest.param to cases that use marks
