### 2023-06-24 9f35388
tests: fix expected message for out-of-order test
