### 2023-06-24 b4cacd3
refactor: change target type to str, fix profile tests
