### 2023-06-24 c7a1e7a
refactor: change target type to str, fix profile tests
