+++ tests/unit/infrastructure/application/ganon/test_ganon_profile.py
@@ -35,7 +35,7 @@ from taxpasta.infrastructure.application import GanonProfile
             OrderedDict(
                 [
                     ("rank", ["root"]),
-                    ("target", [1]),
+                    ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("nr_unique", [1]),
@@ -51,7 +51,7 @@ from taxpasta.infrastructure.application import GanonProfile
                 OrderedDict(
                     [
                         ("rank", ["root"]),
-                        ("target", [1]),
+                        ("target", ["1"]),
                         ("lineage", ["1"]),
                         ("name", ["root"]),
                         ("nr_unique", [1]),
@@ -69,7 +69,7 @@ from taxpasta.infrastructure.application import GanonProfile
             pd.DataFrame(
                 OrderedDict(
                     [
-                        ("target", [1]),
+                        ("target", ["1"]),
                         ("lineage", ["1"]),
                         ("name", ["root"]),
                         ("nr_unique", [1]),
@@ -99,7 +99,7 @@ def test_column_presence(profile: pd.DataFrame):
             OrderedDict(
                 [
                     ("rank", ["root"]),
-                    ("target", [1]),
+                    ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("nr_unique", [1]),
@@ -115,8 +115,8 @@ def test_column_presence(profile: pd.DataFrame):
                 OrderedDict(
                     [
                         ("rank", ["unclassified", "root"]),
-                        ("target", ["-", 1]),
-                        ("lineage", ["-", 1]),
+                        ("target", ["-", "1"]),
+                        ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
                         ("nr_unique", [0, 0]),
                         ("nr_shared", [0, 0]),
@@ -127,7 +127,7 @@ def test_column_presence(profile: pd.DataFrame):
                 )
             ),
             marks=pytest.mark.raises(
-                exception=SchemaError, message="compositionality - percent exceeds 100"
+                exception=SchemaError, message="check_compositionality"
             ),
         ),
         pytest.param(
@@ -135,8 +135,8 @@ def test_column_presence(profile: pd.DataFrame):
                 OrderedDict(
                     [
                         ("rank", ["unclassified", "root"]),
-                        ("target", ["-", 1]),
-                        ("lineage", ["-", 1]),
+                        ("target", ["-", "1"]),
+                        ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
                         ("nr_unique", [0, 0]),
                         ("nr_shared", [0, 0]),
@@ -147,7 +147,7 @@ def test_column_presence(profile: pd.DataFrame):
                 )
             ),
             marks=pytest.mark.raises(
-                exception=SchemaError, message="compositionality - percent below 100"
+                exception=SchemaError, message="check_compositionality"
             ),
         ),
     ],
@@ -164,7 +164,7 @@ def test_percent(profile: pd.DataFrame):
             OrderedDict(
                 [
                     ("rank", ["root"]),
-                    ("target", [1]),
+                    ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("nr_unique", [1]),
@@ -181,38 +181,64 @@ def test_percent(profile: pd.DataFrame):
                     [
                         ("rank", ["unclassified", "root"]),
                         ("target", ["-", 1]),
-                        ("lineage", ["-", 1]),
+                        ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
[…]
