### 2023-06-25 #84 @ctb
Hi @jfy133 sounds good to me - I think the JOSS manuscript is indeed the last thing remaining! I think I can review that all on my own without @snacktavish or @bluegenes so lmk when I should take a look :).
