### 2023-06-27 #84 @lwasser
hi friends! just chiming in here. @ctb you are more than welcome to look at the JOSS manuscript BUT normally that part of this review happens on the JOSS side of things as they need to actually implement the "Accept to JOSS" part. Instructions for wrapping things up (IF this review is at this stage based upon Titus' evaluation of reviews / responses etc) [can be found here!](https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html#how-to-accept-a-package-into-the-pyopensci-ecosystem) . there is a template that you paste into the issue if all of the reviewer elements have been addressed in this review to your satisfaction - then this can move on to JOSS. I hope that helps ✨ i know this partnership can be confusing and sometimes even the JOSS editors need a bit of guidance (ie that they don't have to review the code only the paper through our partnership!).
