### 2023-06-27 #84 @jfy133
I think we meant it's not present a la the checklist in the OP, but indeed the more the merrier in terms of :eyes:!

![image](https://github.com/pyOpenSci/software-submission/assets/17950287/3cccf761-3c56-4e1b-9a36-73627c5b5154)
