### 2023-06-27 #84 @lwasser
ahhh! if i follow you correctly you can absolutely work on the JOSS paper now even if the review is not quite wrapped up. AND i see that note about **not submitting to JOSS separately**. I need to fix that as the process has changed. JOSS now DOES want a submission but you do NOT need to go through another review . You need to link to and mention this review and tell them it was accepted by us. They will and should ONLY review the paper once this review is complete. does that help? NOTE to myself to update the reviewer template and remove that note
