### 2023-07-01 248a99e
fix: address review comments by James and Maxime
