+++ paper/paper.bib
@@ -265,3 +265,34 @@
 	pages = {51},
 	file = {Meyer et al (2019)_Genome Biology.pdf:/home/moritz/Dropbox/References/Storage/Meyer et al (2019)_Genome Biology.pdf:application/pdf},
 }
+
+@article{caporaso_qiime_2010,
+	title = {{QIIME} allows analysis of high-throughput community sequencing data},
+	volume = {7},
+	issn = {1548-7105},
+	url = {https://doi.org/10.1038/nmeth.f.303},
+	doi = {10.1038/nmeth.f.303},
+	number = {5},
+	journal = {Nature Methods},
+	author = {Caporaso, J Gregory and Kuczynski, Justin and Stombaugh, Jesse and Bittinger, Kyle and Bushman, Frederic D and Costello, Elizabeth K and Fierer, Noah and Peña, Antonio Gonzalez and Goodrich, Julia K and Gordon, Jeffrey I and Huttley, Gavin A and Kelley, Scott T and Knights, Dan and Koenig, Jeremy E and Ley, Ruth E and Lozupone, Catherine A and McDonald, Daniel and Muegge, Brian D and Pirrung, Meg and Reeder, Jens and Sevinsky, Joel R and Turnbaugh, Peter J and Walters, William A and Widmann, Jeremy and Yatsunenko, Tanya and Zaneveld, Jesse and Knight, Rob},
+	month = may,
+	year = {2010},
+	pages = {335--336},
+	file = {Caporaso et al (2010)_Nature Methods.pdf:/home/moritz/Dropbox/References/Storage/Caporaso et al (2010)_Nature Methods.pdf:application/pdf},
+}
+
+@book{r_core_team_r_2023,
+	address = {Vienna, Austria},
+	title = {R: {A} {Language} and {Environment} for {Statistical} {Computing}},
+	url = {https://www.R-project.org},
+	publisher = {R Foundation for Statistical Computing},
+	author = {{R Core Team}},
+	year = {2023},
+}
+
+@book{van_rossum_python_1995,
+	title = {Python tutorial},
+	publisher = {Centrum voor Wiskunde en Informatica Amsterdam, The Netherlands},
+	author = {Van Rossum, Guido and Drake Jr, Fred L},
+	year = {1995},
+}
