+++ paper/paper.md
@@ -40,13 +40,13 @@ bibliography: paper.bib
 
 # Summary
 
-Metagenomic analysis is chiefly concerned with untargeted genetic
+Metagenomic analysis is largely concerned with untargeted genetic
 characterisation of the taxonomic and functional composition of whole
 communities of organisms. Researchers ask questions from metagenomic sequencing
 such as 'who is present' (what organisms), and 'what are they doing' (what
 functions are they performing)? The nature of this field is such that it
 intersects with ecology, medicine, statistics, and bioinformatics. Facilitated
-by the development of Next-Generation-Sequencing (NGS), the field often
+by the development of Next-Generation Sequencing (NGS), the field often
 generates large datasets consisting of many samples (hundreds) and many
 sequencing reads (tens of millions).
 
@@ -61,12 +61,12 @@ metagenomic profiles. However, most of those tools have invented their own
 (often tabular) result formats, which complicates downstream analysis and in
 particular comparison across tools.
 
-TAXPASTA is a standalone command-line tool written in Python that aims to
-standardise the diverse range of metagenomic profiler output formats to simple
-tabular formats that are readily consumed in downstream applications. TAXPASTA
-facilitates cross-comparison between taxonomic profiling tools without the need
-for external or dedicated modules or plugins needed of other 'dedicated'
-metagenomic profile formats.
+TAXPASTA is a standalone command-line tool written in Python
+[@van_rossum_python_1995] that aims to standardise the diverse range of
+metagenomic profiler output formats to simple tabular formats that are readily
+consumed in downstream applications. TAXPASTA facilitates cross-comparison
+between taxonomic profiling tools without the need for external or dedicated
+modules or plugins needed of other 'dedicated' metagenomic profile formats.
 
 # Statement of need
 
@@ -77,13 +77,13 @@ nf-core community [@ewels_nf-core_2020].
 
 Across profilers, relative abundances can be reported in read counts, fractions,
 or percentages, as well as any number of additional columns with extra
-information. Taxonomic identifiers can be recorded with taxonmic identifiers,
-taxonomic names and/or in some cases semi-colon-separated taxonomic 'paths'
-(lineages). These can also be formatted in different ways, from typical tables,
-to including 'indented' taxonomy trees such as in the Kraken
-[@wood_improved_2019] family of profilers. Manually parsing these for comparison
-can be an arduous, error-prone task, with researchers often reverting to custom
-R and Python scripting, or even manual correction in spreadsheet software.
+information. Taxa can be recorded using taxonomic identifiers, taxonomic names
+and/or in some cases semi-colon-separated taxonomic 'paths' (lineages). These
+can also be formatted in different ways, from typical tables, to including
+'indented' taxonomy trees such as in the Kraken [@wood_improved_2019] family of
+profilers. Manually parsing these for comparison can be an arduous, error-prone
+task, with researchers often reverting to custom R [@r_core_team_r_2023] and
+Python scripting, or even manual correction in spreadsheet software.
 
 With TAXPASTA, all of those formats can be converted into a single, standardised
 output, that, at a minimum, contains taxonomic identifiers and their relative
@@ -95,34 +95,36 @@ but it is by far the most comprehensive in terms of supported profilers and
 output formats.
 
 There exists an initiative to benchmark and compare profilers, as well as
-provide guidance on their fitness for purpose
-[@sczyrba_critical_2017; @meyer_critical_2022]. For that purpose, the
-Open-community Profiling Assessment tooL (OPAL) [@meyer_assessing_2019] was
-developed. Creating a community wide assessment faced many of the challenges
[…]
