### 2023-07-01 #84 @Midnighter
Dear @ctb,

You can have a look at our [draft for JOSS](https://github.com/taxprofiler/taxpasta/suites/13998480996/artifacts/781056400) if you like, but as @lwasser mentioned this is not required to complete the review process here. Either way, we look forward to your final comments.
