### 2023-07-01 1e9c73c
Minor grammartical and phrasing fixes
