### 2023-07-02 fbe1346
fix: small wording changes
