+++ paper/paper.md
@@ -44,7 +44,7 @@ Metagenomic analysis is largely concerned with untargeted genetic
 characterisation of the taxonomic and functional composition of whole
 communities of organisms. Researchers ask questions from metagenomic sequencing
 such as 'who is present' (what organisms are present), and 'what are they doing'
-(what functions are they performing)? The nature of this field is such that it
+(which functions are they performing)? The nature of this field is such that it
 intersects with ecology, medicine, statistics, and bioinformatics. Facilitated
 by the development of Next-Generation Sequencing (NGS), the field often
 generates large datasets consisting of many samples (hundreds) and many
@@ -89,7 +89,7 @@ Python scripting, or even manual correction in spreadsheet software.
 
 With TAXPASTA, all of those formats can be converted into a single, standardised
 output, that, at a minimum, contains taxonomic identifiers and their relative
-sequence abundances as integer counts. It can also be used to aggregate profiles across
+abundances as integer counts. It can also be used to aggregate profiles across
 samples from the same profiler and merge them into a single, standardised table.
 Having a singular format facilitates downstream analyses and comparisons.
 TAXPASTA is not the first tool to attempt standardising metagenomic profiles,
@@ -121,7 +121,7 @@ some of the taxonomic profilers also come with scripts to convert their output
 into another format and/or merge multiple profiles into a single table, such as
 the Krakentools companion package [@lu_metagenome_2022], these are often focused
 on the specific tool or family of tools. Thus, users would have to become
-proficient in yet another piece of software per tool or family of tools just for 
+proficient in yet another piece of software per tool or family of tools for the sake of
 consistent output files.
 
 [^8]: <https://github.com/bioboxes/rfc/tree/master/data-format>
