+++ tests/unit/infrastructure/application/ganon/test_ganon_profile.py
@@ -38,7 +38,7 @@ from taxpasta.infrastructure.application import GanonProfile
                     ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
-                    ("nr_unique", [1]),
+                    ("number_unique", [1]),
                     ("nr_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
@@ -54,7 +54,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("target", ["1"]),
                         ("lineage", ["1"]),
                         ("name", ["root"]),
-                        ("nr_unique", [1]),
+                        ("number_unique", [1]),
                         ("nr_shared", [1]),
                         ("nr_children", [1]),
                         ("nr_cumulative", [1]),
@@ -72,7 +72,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("target", ["1"]),
                         ("lineage", ["1"]),
                         ("name", ["root"]),
-                        ("nr_unique", [1]),
+                        ("number_unique", [1]),
                         ("nr_shared", [1]),
                         ("nr_children", [1]),
                         ("nr_cumulative", [1]),
@@ -102,7 +102,7 @@ def test_column_presence(profile: pd.DataFrame):
                     ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
-                    ("nr_unique", [1]),
+                    ("number_unique", [1]),
                     ("nr_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
@@ -118,7 +118,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("target", ["-", "1"]),
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
-                        ("nr_unique", [0, 0]),
+                        ("number_unique", [0, 0]),
                         ("nr_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
@@ -138,7 +138,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("target", ["-", "1"]),
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
-                        ("nr_unique", [0, 0]),
+                        ("number_unique", [0, 0]),
                         ("nr_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
@@ -167,7 +167,7 @@ def test_percent(profile: pd.DataFrame):
                     ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
-                    ("nr_unique", [1]),
+                    ("number_unique", [1]),
                     ("nr_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
@@ -183,7 +183,7 @@ def test_percent(profile: pd.DataFrame):
                         ("target", ["-", 1]),
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
-                        ("nr_unique", [0, -1]),
+                        ("number_unique", [0, -1]),
                         ("nr_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
@@ -213,7 +213,7 @@ def test_target(profile: pd.DataFrame):
                     ("target", ["1"]),
                     ("lineage", ["1"]),
                     ("name", ["root"]),
-                    ("nr_unique", [1]),
[…]
