+++ src/taxpasta/infrastructure/application/ganon/ganon_profile.py
@@ -35,7 +35,7 @@ class GanonProfile(BaseDataFrameModel):
     lineage: Series[str] = pa.Field()
     name: Series[str] = pa.Field()
     number_unique: Series[int] = pa.Field(ge=0)
-    nr_shared: Series[int] = pa.Field(ge=0)
+    number_shared: Series[int] = pa.Field(ge=0)
     nr_children: Series[int] = pa.Field(ge=0)
     nr_cumulative: Series[int] = pa.Field(ge=0)
     pc_cumulative: Series[float] = pa.Field(ge=0.0, le=100.0)
