+++ tests/unit/infrastructure/application/ganon/test_ganon_profile.py
@@ -39,7 +39,7 @@ from taxpasta.infrastructure.application import GanonProfile
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("number_unique", [1]),
-                    ("nr_shared", [1]),
+                    ("number_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
@@ -55,7 +55,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("lineage", ["1"]),
                         ("name", ["root"]),
                         ("number_unique", [1]),
-                        ("nr_shared", [1]),
+                        ("number_shared", [1]),
                         ("nr_children", [1]),
                         ("nr_cumulative", [1]),
                     ]
@@ -73,7 +73,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("lineage", ["1"]),
                         ("name", ["root"]),
                         ("number_unique", [1]),
-                        ("nr_shared", [1]),
+                        ("number_shared", [1]),
                         ("nr_children", [1]),
                         ("nr_cumulative", [1]),
                         ("pc_cumulative", [100.0]),
@@ -103,7 +103,7 @@ def test_column_presence(profile: pd.DataFrame):
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("number_unique", [1]),
-                    ("nr_shared", [1]),
+                    ("number_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
@@ -119,7 +119,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
                         ("number_unique", [0, 0]),
-                        ("nr_shared", [0, 0]),
+                        ("number_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 39.61288]),
@@ -139,7 +139,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
                         ("number_unique", [0, 0]),
-                        ("nr_shared", [0, 0]),
+                        ("number_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 9.61288]),
@@ -168,7 +168,7 @@ def test_percent(profile: pd.DataFrame):
                     ("lineage", ["1"]),
                     ("name", ["root"]),
                     ("number_unique", [1]),
-                    ("nr_shared", [1]),
+                    ("number_shared", [1]),
                     ("nr_children", [1]),
                     ("nr_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
@@ -184,7 +184,7 @@ def test_percent(profile: pd.DataFrame):
                         ("lineage", ["-", "1"]),
                         ("name", ["unclassified", "root"]),
                         ("number_unique", [0, -1]),
-                        ("nr_shared", [0, 0]),
+                        ("number_shared", [0, 0]),
                         ("nr_children", [0, 457530]),
                         ("nr_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 27.61288]),
@@ -214,7 +214,7 @@ def test_target(profile: pd.DataFrame):
                     ("lineage", ["1"]),
                     ("name", ["root"]),
[…]
