### 2023-07-02 4a7daf0
refactor: rename number of cumulative field
