+++ tests/unit/infrastructure/application/ganon/test_ganon_profile.py
@@ -41,7 +41,7 @@ from taxpasta.infrastructure.application import GanonProfile
                     ("number_unique", [1]),
                     ("number_shared", [1]),
                     ("number_children", [1]),
-                    ("nr_cumulative", [1]),
+                    ("number_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
                 ]
             )
@@ -57,7 +57,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("number_unique", [1]),
                         ("number_shared", [1]),
                         ("number_children", [1]),
-                        ("nr_cumulative", [1]),
+                        ("number_cumulative", [1]),
                     ]
                 )
             ),
@@ -75,7 +75,7 @@ from taxpasta.infrastructure.application import GanonProfile
                         ("number_unique", [1]),
                         ("number_shared", [1]),
                         ("number_children", [1]),
-                        ("nr_cumulative", [1]),
+                        ("number_cumulative", [1]),
                         ("pc_cumulative", [100.0]),
                         ("rank", ["root"]),
                     ]
@@ -105,7 +105,7 @@ def test_column_presence(profile: pd.DataFrame):
                     ("number_unique", [1]),
                     ("number_shared", [1]),
                     ("number_children", [1]),
-                    ("nr_cumulative", [1]),
+                    ("number_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
                 ]
             )
@@ -121,7 +121,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("number_unique", [0, 0]),
                         ("number_shared", [0, 0]),
                         ("number_children", [0, 457530]),
-                        ("nr_cumulative", [0, 4575301]),
+                        ("number_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 39.61288]),
                     ]
                 )
@@ -141,7 +141,7 @@ def test_column_presence(profile: pd.DataFrame):
                         ("number_unique", [0, 0]),
                         ("number_shared", [0, 0]),
                         ("number_children", [0, 457530]),
-                        ("nr_cumulative", [0, 4575301]),
+                        ("number_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 9.61288]),
                     ]
                 )
@@ -170,7 +170,7 @@ def test_percent(profile: pd.DataFrame):
                     ("number_unique", [1]),
                     ("number_shared", [1]),
                     ("number_children", [1]),
-                    ("nr_cumulative", [1]),
+                    ("number_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
                 ]
             )
@@ -186,7 +186,7 @@ def test_percent(profile: pd.DataFrame):
                         ("number_unique", [0, -1]),
                         ("number_shared", [0, 0]),
                         ("number_children", [0, 457530]),
-                        ("nr_cumulative", [0, 4575301]),
+                        ("number_cumulative", [0, 4575301]),
                         ("pc_cumulative", [72.38712, 27.61288]),
                     ]
                 )
@@ -216,7 +216,7 @@ def test_target(profile: pd.DataFrame):
                     ("number_unique", [1]),
                     ("number_shared", [1]),
                     ("number_children", [1]),
-                    ("nr_cumulative", [1]),
+                    ("number_cumulative", [1]),
                     ("pc_cumulative", [100.0]),
                 ]
             )
@@ -232,7 +232,7 @@ def test_target(profile: pd.DataFrame):
                         ("number_unique", [0, -1]),
                         ("number_shared", [0, 0]),
[…]
