### 2023-07-02 9dcaaed
refactor: rename number of cumulative field
