### 2023-07-02 77a0b24
refactor: rename percent cumulative field
