### 2023-07-02 4265627
refactor: rename percent cumulative field
