### 2023-07-02 963a6e4
fix: convert taxonomy IDs to integers
