### 2023-07-02 ef280dc
refactor: move header tests into methods
Handle buffer arguments in addition to path-like arguments.
