### 2023-07-02 7349f16
refactor: move header tests into methods
Handle buffer arguments in addition to path-like arguments.
