### 2023-07-02 a257bfb
docs: prepare changelog for release 0.4.0
