### 2023-07-02 #84 @Midnighter
FYI, we just released version 0.4.0 that wraps the changes from the reviews, as well as further issues discovered by users.
