### 2023-07-03 d7d1835
Add logos, favicon, and diagram
