### 2023-07-03 b33272f
Add logos, favicon, and diagram
