### 2023-07-04 6343a40
docs: move "About" section and name the page "Home"
