### 2023-07-04 618fcdc
docs: insert direct link to badge
