### 2023-07-04 c00494c
docs: remove short description line
