### 2023-07-04 7c63700
docs: add link to ganon to the readme
