### 2023-07-05 e5d1631
docs: add a white logo for the header
