### 2023-07-05 2278ace
docs: add summary figure
