### 2023-07-05 aaa0040
chore: add ganon reference
