### 2023-07-06 #84 @Kevin-Mattheus-Moerman
@ctb @lwasser :wave: I am an editor for JOSS and processing the JOSS side of things. Can you clarify if this review is fully complete and this submission has passed review? The reason I ask is that many of the boxes :point_up: are not ticked.
