### 2023-07-07 #84 @Kevin-Mattheus-Moerman
@ctb thanks for the reply. Note though that @NickleDave "Editor in Chief checks" set has unchecked boxes :point_up:, and so does your `Author Wrap Up Tasks` and `Editor Final Checks` set. It seems to me like most of these can just be ticked? If so I think it would be best if your group makes a habit of ticking them (perhaps you still can now), so that somebody studying this review issue is not confused about the software/review lacking anything.
