### 2023-07-07 #84 @ctb
> @ctb thanks for the reply. Note though that @NickleDave `Editor in Chief checks` set has unchecked boxes ☝️, and so does your `Author Wrap Up Tasks` and `Editor Final Checks` set. It seems to me like most of these can just be ticked? If so I think it would be best if your group makes a habit of ticking them (perhaps you still can now), so that somebody studying this review issue is not confused about the software/review lacking anything.

there are _so many_ check boxes 😆

but also, I think:

>Invite the maintainers to submit a blog post highlighting their package.

could maybe best happen after JOSS paper is available. Similarly,

>If the author submits to JOSS...

I felt needed to wait until there was a JOSS issue. In addition, I don't think the authors have the ability to check off those boxes so it had to wait until THEY did the work and then I noticed that they had done so => check the boxes.

For the other checklist items like adding badges and so on, those were done _after_ I signed off on the review and I don't get notified of PRs etc. referencing this issue so barring monitoring this issue manually or being tagged in on the PRs, I'm not sure how I would go about knowing I should check them off...

(I don't mean to be defensive - I am just explaining where we are and why, and confirming to myself that I had read the unchecked boxes and made the decision not to check them off yet, instead of missing them entirely, which was also certainly a possibility!)
