### 2023-07-07 #84 @jfy133
> In addition, I don't think the authors have the ability to check off those boxes so it had to wait until THEY did the work and then I noticed that they had done so => check the boxes.

Correct, at least I was unable to tick of the boxes of what was meant to be our tasks (as authors)
