### 2023-07-11 6407ec1
chore: add suggestions from JOSS
[skip ci]
