+++ CITATION.cff
@@ -1,46 +1,69 @@
-# This CITATION.cff file was generated with cffinit.
-# Visit https://bit.ly/cffinit to generate yours today!
-
-cff-version: 1.2.0
+cff-version: "1.2.0"
 title: >-
   TAXPASTA: TAXonomic Profile Aggregation and
   STAndardisation
 message: >-
-  If you use this software, please cite it using the
-  metadata from this file.
-type: software
+  If you use this software, please cite our article in the
+  Journal of Open Source Software.
+doi: 10.5281/zenodo.8105840
 authors:
   - given-names: Moritz E.
     family-names: Beber
-    affiliation: 'Unseen Bio ApS, Copenhagen, Denmark'
-    orcid: 'https://orcid.org/0000-0003-2406-1978'
+    affiliation: "Unseen Bio ApS, Copenhagen, Denmark"
+    orcid: "https://orcid.org/0000-0003-2406-1978"
   - given-names: Maxime
     family-names: Borry
     affiliation: >-
       Microbiome Sciences Group, Department of
       Archaeogenetics, Max Planck Institute for Evolutionary
       Anthropology, Leipzig, Germany
-    orcid: 'https://orcid.org/0000-0001-9140-7559'
+    orcid: "https://orcid.org/0000-0001-9140-7559"
   - given-names: Sofia
     family-names: Stamouli
-    orcid: 'https://orcid.org/0009-0006-0893-3771'
+    orcid: "https://orcid.org/0009-0006-0893-3771"
     affiliation: >-
       Department of Microbiology, Tumor and Cell Biology,
       Karolinska Institute, Solna, Sweden
   - given-names: James A.
     family-names: Fellows Yates
-    orcid: 'https://orcid.org/0000-0001-5585-6277'
+    orcid: "https://orcid.org/0000-0001-5585-6277"
     affiliation: >-
       Microbiome Sciences Group, Department of
       Archaeogenetics, Max Planck Institute for Evolutionary
       Anthropology, Leipzig, Germany
-identifiers:
-  - type: doi
-    value: 10.5281/zenodo.8105840
-    description: Zenodo archive of published version 0.4.0.
-repository-code: 'https://github.com/taxprofiler/taxpasta'
-url: 'https://taxpasta.readthedocs.io/'
-repository-artifact: 'https://zenodo.org/record/8105840'
+contact:
+- family-names: Beber
+  given-names: Moritz E.
+  orcid: "https://orcid.org/0000-0003-2406-1978"
+repository-code: "https://github.com/taxprofiler/taxpasta"
+url: "https://taxpasta.readthedocs.io/"
+repository-artifact: "https://zenodo.org/record/8105840"
+preferred-citation:
+  authors:
+  - family-names: Beber
+    given-names: Moritz E.
+    orcid: "https://orcid.org/0000-0003-2406-1978"
+  - family-names: Borry
+    given-names: Maxime
+    orcid: "https://orcid.org/0000-0001-9140-7559"
+  - family-names: Stamouli
+    given-names: Sofia
+    orcid: "https://orcid.org/0009-0006-0893-3771"
+  - family-names: Fellows Yates
+    given-names: James A.
+    orcid: "https://orcid.org/0000-0001-5585-6277"
+  date-published: 2023-07-11
+  doi: 10.21105/joss.05627
+  issn: 2475-9066
+  issue: 87
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5627
+  title: "TAXPASTA: TAXonomic Profile Aggregation and STAndardisation"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05627"
+  volume: 8
 keywords:
   - bioinformatics
   - metagenomics
