### 2023-07-12 #84 @lwasser
ahhhh ok lessons learned here that we need a way for authors to check their boxes AND/OR we need fewer of them . thank you all for the feedback!!! i hadn't considered that authors would NOT be able to check those author list boxes - we will have to come up with a better approach! i'm open to suggestions.

and it looks like  [taxpasta was accepted by JOSS](https://github.com/openjournals/joss-reviews/issues/5627) - hooray. 🎉 🎉
