### 2023-07-12 #84 @lwasser
awesome! ok i'll close this issue now. Friends - if anyone from this review (reviewers, maintainers etc) is interested in joining our slack community please email me at leah at pyopensci.org  or reply here if your email is available on your github and i can add you!! we'd love to have all of you there!
